ci: assemble checksummed P0/P1 patch units
