+++ scripts/p0p1_chunks/app_main.part.00
@@ -1,0 +1,1 @@
+ZnJvbSBfX2Z1dHVyZV9fIGltcG9ydCBhbm5vdGF0aW9ucwoKaW1wb3J0IHJlCmZyb20gcGF0aGxpYiBpbXBvcnQgUGF0aApmcm9tIHRleHR3cmFwIGltcG9ydCBkZWRlbnQKClJPT1QgPSBQYXRoLmN3ZCgpCgoKZGVmIHJlYWQocGF0aDogc3RyKSAtPiBzdHI6CiAgICByZXR1cm4gKFJPT1QgLyBwYXRoKS5yZWFkX3RleHQoZW5jb2Rpbmc9InV0Zi04IikKCgpkZWYgd3JpdGUocGF0aDogc3RyLCB0ZXh0OiBzdHIpIC0+IE5vbmU6CiAgICB0YXJnZXQgPSBST09UIC8gcGF0aAogICAgdGFyZ2V0LnBhcmVudC5ta2RpcihwYXJlbnRzPVRydWUsIGV4aXN0X29rPVRydWUpCiAgICB0YXJnZXQud3JpdGVfdGV4dCh0ZXh0LCBlbmNvZGluZz0idXRmLTgiKQoKCmRlZiByZXBsYWNlX29uY2UodGV4dDogc3RyLCBvbGQ6IHN0ciwgbmV3OiBzdHIsICosIGxhYmVsOiBzdHIpIC0+IHN0cjoKICAgIGNvdW50ID0gdGV4dC5jb3VudChvbGQpCiAgICBpZiBjb3VudCAhPSAxOgogICAgICAgIHJhaXNlIFJ1bnRpbWVFcnJvcihmIntsYWJlbH06IGV4cGVjdGVkIGV4YWN0bHkgb25lIG1hdGNoLCBmb3VuZCB7Y291bnR9IikKICAgIHJldHVybiB0ZXh0LnJlcGxhY2Uob2xkLCBuZXcsIDEpCgoKZGVmIHJlZ2V4X29uY2UodGV4dDogc3RyLCBwYXR0ZXJuOiBzdHIsIHJlcGxhY2VtZW50OiBzdHIsICosIGxhYmVsOiBzdHIsIGZsYWdzOiBpbnQgPSAwKSAtPiBzdHI6CiAgICB1cGRhdGVkLCBjb3VudCA9IHJlLnN1Ym4ocGF0dGVybiwgcmVwbGFjZW1lbnQsIHRleHQsIGNvdW50PTEsIGZsYWdzPWZsYWdzKQogICAgaWYgY291bnQgIT0gMToKICAgICAgICByYWlzZSBSdW50aW1lRXJyb3IoZiJ7bGFiZWx9OiBleHBlY3RlZCBleGFjdGx5IG9uZSByZWdleCBtYXRjaCwgZm91bmQge2NvdW50fSIpCiAgICByZXR1cm4gdXBkYXRlZAoKCiMgLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tCiMgQXBwIHdpcmluZzogcmVtb3ZlIGR1cGxpY2F0ZSBpbmxpbmUgd2F0Y2hkb2cgYW5kIGdhdGUgbGl2ZSBhcm1pbmcgb24gU0xPIHN0YXRlLgojIC0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLQpwYXRoID0gInNyYy9uaWZ0eV9zY2FscGVyX2JvdC9jb3JlL2FwcC5weSIKdGV4dCA9IHJlYWQocGF0aCkKdGV4dCA9IHJlZ2V4X29uY2UoCiAgICB0ZXh0LAogICAgciJcbiAgICAgICAgIyBcW0ZJWFxdIFN0YXJ0IEhlYWx0aCBNb25pdG9yIFwoV2F0Y2hkb2dcKSBmb3IgUG9sbGluZyBNb2RlXG4uKj9cbiAgICAgICAgZGF0YV9odWIgPSBEYXRhSHViXCgiLAogICAgIiIiXG4gICAgICAgICMgT25lIG1hcmtldC1hd2FyZSBzdXBlcnZpc29yIG93bnMgcmVzdGFydCwgZmF0YWwtc3RhbGUgYW5kIGxhdGVuY3kgcG9saWN5LlxuICAgICAgICB0cnk6XG4gICAgICAgICAgICBzdGFydF93YXRjaGRvZyhtYXJrZXRfZGF0YV9tYW5hZ2VyKVxuICAgICAgICAgICAgTE9HR0VSLmluZm8oXCJEYXRhIEhlYWx0aCBNb25pdG9yIChzaW5nbGUgc3VwZXJ2aXNvcikgc3RhcnRlZFwiKVxuICAgICAgICBleGNlcHQgRXhjZXB0aW9uIGFzIGV4YzpcbiAgICAgICAgICAgIExPR0dFUi53YXJuaW5nKFwiRmFpbGVkIHRvIHN0YXJ0IG1hcmtldC1kYXRhIHN1cGVydmlzb3I6ICVzXCIsIGV4YylcblxuICAgICAgICBkYXRhX2h1YiA9IERhdGFIdWIoIiIiLAogICAgbGFiZWw9InJlbW92ZSBkdXBsaWNhdGUgd2F0Y2hkb2ciLAogICAgZmxhZ3M9cmUuUywKKQp0ZXh0ID0gcmVwbGFjZV9vbmNlKAogICAgdGV4dCwKICAgICIiIiAgICBleGVjdXRpb25fY2FwYWNpdHlfcmVhZHk6IGJvb2wgPSBGYWxzZVxuICAgIG1pbmltdW1fbG90X2FmZm9yZGFiaWxpdHlfYnlfc3ltYm9sOiBkaWN0W3N0ciwgZGljdFtzdHIsIG9iamVjdF1dID0gZmllbGQoXG4iIiIsCiAgICAiIiIgICAgZXhlY3V0aW9uX2NhcGFjaXR5X3JlYWR5OiBib29sID0gRmFsc2VcbiAgICBtYXJrZXRfZGF0YV9sYXRlbmN5X2RlZ3JhZGVkOiBib29sID0gRmFsc2VcbiAgICBtaW5pbXVtX2xvdF9hZmZvcmRhYmlsaXR5X2J5X3N5bWJvbDogZGljdFtzdHIsIGRpY3Rbc3RyLCBvYmplY3RdXSA9IGZpZWxkKFxuIiIiLAogICAgbGFiZWw9ImNvbnRleHQgbGF0ZW5jeSBmaWVsZCIsCikKdGV4dCA9IHJlcGxhY2Vfb25jZSgKICAgIHRleHQsCiAgICAiIiIgICAgbGl2ZV9vcmRlcnNfYXJtZWQgPSBib29sKFxuICAgICAgICBsaXZlX21vZGVcbiAgICAgICAgYW5kIG1hcmtldF9vcGVuXG4gICAgICAgIGFuZCBldmFsdWF0aW9uX3JlYWR5XG4gICAgICAgIGFuZCBjb250ZXh0X2V4ZWNfcmVhZHlcbiAgICAgICAgYW5kIGJyb2tlcl9yZWFkeVxuICAgICAgICBhbmQgYWxsX3NlbGVjdGVkX29wdGlvbnNfZXhlY19yZWFkeVxuICAgIClcbiIiIiwKICAgICIiIiAgICBsYXRlbmN5X2RlZ3JhZGVkID0gYm9vbChcbiAgICAgICAgZ2V0YXR0cihnZXRhdHRyKGN0eCwgXCJtYXJrZXRfZGF0YV9tYW5hZ2VyXCIsIE5vbmUpLCBcImxhdGVuY3lfZGVncmFkZWRcIiwgRmFsc2UpXG4gICAgKVxuICAgIGxpdmVfb3JkZXJzX2FybWVkID0gYm9vbChcbiAgICAgICAgbGl2ZV9tb2RlXG4gICAgICAgIGFuZCBtYXJrZXRfb3BlblxuICAgICAgICBhbmQgZXZhbHVhdGlvbl9yZWFkeVxuICAgICAgICBhbmQgY29udGV4dF9leGVjX3JlYWR5XG4gICAgICAgIGFuZCBicm9rZXJfcmVhZHlcbiAgICAgICAgYW5kIGFsbF9zZWxlY3RlZF9vcHRpb25zX2V4ZWNfcmVhZHlcbiAgICAgICAgYW5kIG5vdCBsYXRlbmN5X2RlZ3JhZGVkXG4gICAgKVxuIiIiLAogICAgbGFiZWw9ImxhdGVuY3kgYXJtaW5nIGdhdGUiLAopCnRleHQgPSByZXBsYWNlX29uY2UoCiAgICB0ZXh0LAogICAgIiIiICAgIGlmIG5vdCBzZWxlY3RlZF9jZTpcbiAgICAgICAgbWlzc2luZy5hcHBlbmQoXCJzZWxlY3RlZF9jZV9taXNzaW5nXCIpXG4iIiIsCiAgICAiIiIgICAgaWYgbGl2ZV9tb2RlIGFuZCBsYXRlbmN5X2RlZ3JhZGVkOlxuICAgICAgICBtaXNzaW5nLmFwcGVuZChcIm1hcmtldF9kYXRhX2xhdGVuY3lfZGVncmFkZWRcIilcbiAgICBpZiBub3Qgc2VsZWN0ZWRfY2U6XG4gICAgICAgIG1pc3NpbmcuYXBwZW5kKFwic2VsZWN0ZWRfY2VfbWlzc2luZ1wiKVxuIiIiLAogICAgbGFiZWw9ImxhdGVuY3kgYmxvY2tlciIsCikKdGV4dCA9IHJlcGxhY2Vfb25jZSgKICAgIHRleHQsCiAgICAiIiIgICAgY3R4LmV4ZWN1dGlvbl9jYXBhY2l0eV9yZWFkeSA9IGJvb2woZXhlY3V0aW9uX2NhcGFjaXR5X3JlYWR5KVxuICAgIGN0eC5taW5pbXVtX2xvdF9hZmZvcmRhYmlsaXR5X2J5X3N5bWJvbCA9IGRpY3QoYWZmb3JkYWJpbGl0eV9ieV9zeW1ib2wpXG4iIiIsCiAgICAiIiIgICAgY3R4LmV4ZWN1dGlvbl9jYXBhY2l0eV9yZWFkeSA9IGJvb2woZXhlY3V0aW9uX2NhcGFjaXR5X3JlYWR5KVxuICAgIGN0eC5tYXJrZXRfZGF0YV9sYXRlbmN5X2RlZ3JhZGVkID0gYm9vbChsYXRlbmN5X2RlZ3JhZGVkKVxuICAgIGN0eC5taW5pbXVtX2xvdF9hZmZvcmRhYmlsaXR5X2J5X3N5bWJvbCA9IGRpY3QoYWZmb3JkYWJpbGl0eV9ieV9zeW1ib2wpXG4iIiIsCiAgICBsYWJlbD0ibGF0ZW5jeSBjb250ZXh0IGFzc2lnbm1lbnQiLAopCndyaXRlKHBhdGgsIHRleHQpCgojIC0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLQojIE1haW4g
