+++ scripts/p0p1_chunks/app_main.part.01
@@ -1,0 +1,1 @@
+cHJvY2VzcyBwdWJsaXNoZXMgZXZlbnQtbG9vcCBsYWcgYW5kIHJlYWR5eiBleHBvc2VzIGxhdGVuY3kgYmxvY2tlci4KIyAtLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0KcGF0aCA9ICJzcmMvbmlmdHlfc2NhbHBlcl9ib3QvbWFpbi5weSIKdGV4dCA9IHJlYWQocGF0aCkKdGV4dCA9IHJlcGxhY2Vfb25jZSgKICAgIHRleHQsCiAgICAiIiIgICAgICAgIF9FVkVOVF9MT09QX0xBR19NUyA9IGxhZ19tc1xuICAgICAgICBleHBlY3RlZCA9IG5vdyArIGludGVydmFsX3NcbiIiIiwKICAgICIiIiAgICAgICAgX0VWRU5UX0xPT1BfTEFHX01TID0gbGFnX21zXG4gICAgICAgIHRyeTpcbiAgICAgICAgICAgIGN0eCA9IF9sYXRlc3RfY29udGV4dCgpXG4gICAgICAgICAgICBtZG0gPSBnZXRhdHRyKGN0eCwgXCJtYXJrZXRfZGF0YV9tYW5hZ2VyXCIsIE5vbmUpIGlmIGN0eCBpcyBub3QgTm9uZSBlbHNlIE5vbmVcbiAgICAgICAgICAgIGlmIG1kbSBpcyBub3QgTm9uZTpcbiAgICAgICAgICAgICAgICBtZG0uX2V2ZW50X2xvb3BfbGFnX21zID0gbGFnX21zXG4gICAgICAgIGV4Y2VwdCBFeGNlcHRpb246XG4gICAgICAgICAgICBwYXNzXG4gICAgICAgIGV4cGVjdGVkID0gbm93ICsgaW50ZXJ2YWxfc1xuIiIiLAogICAgbGFiZWw9InB1Ymxpc2ggZXZlbnQtbG9vcCBsYWciLAopCnRleHQgPSByZXBsYWNlX29uY2UoCiAgICB0ZXh0LAogICAgIiIiICAgIGlmIGJvb2woZ2V0YXR0cihjdHgsIFwicG9zaXRpb25fcmVjb25jaWxpYXRpb25fZmFpbGVkXCIsIEZhbHNlKSk6XG4gICAgICAgIGJsb2NrZXJzLmFwcGVuZChcInBvc2l0aW9uX3JlY29uY2lsaWF0aW9uX2ZhaWxlZFwiKVxuIiIiLAogICAgIiIiICAgIGlmIGJvb2woZ2V0YXR0cihjdHgsIFwicG9zaXRpb25fcmVjb25jaWxpYXRpb25fZmFpbGVkXCIsIEZhbHNlKSk6XG4gICAgICAgIGJsb2NrZXJzLmFwcGVuZChcInBvc2l0aW9uX3JlY29uY2lsaWF0aW9uX2ZhaWxlZFwiKVxuICAgIG1kbSA9IGdldGF0dHIoY3R4LCBcIm1hcmtldF9kYXRhX21hbmFnZXJcIiwgTm9uZSlcbiAgICBpZiBib29sKFxuICAgICAgICBnZXRhdHRyKGN0eCwgXCJtYXJrZXRfZGF0YV9sYXRlbmN5X2RlZ3JhZGVkXCIsIEZhbHNlKVxuICAgICAgICBvciBnZXRhdHRyKG1kbSwgXCJsYXRlbmN5X2RlZ3JhZGVkXCIsIEZhbHNlKVxuICAgICk6XG4gICAgICAgIGJsb2NrZXJzLmFwcGVuZChcIm1hcmtldF9kYXRhX2xhdGVuY3lfZGVncmFkZWRcIilcbiIiIiwKICAgIGxhYmVsPSJyZWFkeXogbGF0ZW5jeSBibG9ja2VyIiwKKQp3cml0ZShwYXRoLCB0ZXh0KQoK
