+++ scripts/p0p1_chunks/tests.part.00
@@ -1,0 +1,1 @@
+ZnJvbSBfX2Z1dHVyZV9fIGltcG9ydCBhbm5vdGF0aW9ucwoKaW1wb3J0IHJlCmZyb20gcGF0aGxpYiBpbXBvcnQgUGF0aApmcm9tIHRleHR3cmFwIGltcG9ydCBkZWRlbnQKClJPT1QgPSBQYXRoLmN3ZCgpCgoKZGVmIHJlYWQocGF0aDogc3RyKSAtPiBzdHI6CiAgICByZXR1cm4gKFJPT1QgLyBwYXRoKS5yZWFkX3RleHQoZW5jb2Rpbmc9InV0Zi04IikKCgpkZWYgd3JpdGUocGF0aDogc3RyLCB0ZXh0OiBzdHIpIC0+IE5vbmU6CiAgICB0YXJnZXQgPSBST09UIC8gcGF0aAogICAgdGFyZ2V0LnBhcmVudC5ta2RpcihwYXJlbnRzPVRydWUsIGV4aXN0X29rPVRydWUpCiAgICB0YXJnZXQud3JpdGVfdGV4dCh0ZXh0LCBlbmNvZGluZz0idXRmLTgiKQoKCmRlZiByZXBsYWNlX29uY2UodGV4dDogc3RyLCBvbGQ6IHN0ciwgbmV3OiBzdHIsICosIGxhYmVsOiBzdHIpIC0+IHN0cjoKICAgIGNvdW50ID0gdGV4dC5jb3VudChvbGQpCiAgICBpZiBjb3VudCAhPSAxOgogICAgICAgIHJhaXNlIFJ1bnRpbWVFcnJvcihmIntsYWJlbH06IGV4cGVjdGVkIGV4YWN0bHkgb25lIG1hdGNoLCBmb3VuZCB7Y291bnR9IikKICAgIHJldHVybiB0ZXh0LnJlcGxhY2Uob2xkLCBuZXcsIDEpCgoKZGVmIHJlZ2V4X29uY2UodGV4dDogc3RyLCBwYXR0ZXJuOiBzdHIsIHJlcGxhY2VtZW50OiBzdHIsICosIGxhYmVsOiBzdHIsIGZsYWdzOiBpbnQgPSAwKSAtPiBzdHI6CiAgICB1cGRhdGVkLCBjb3VudCA9IHJlLnN1Ym4ocGF0dGVybiwgcmVwbGFjZW1lbnQsIHRleHQsIGNvdW50PTEsIGZsYWdzPWZsYWdzKQogICAgaWYgY291bnQgIT0gMToKICAgICAgICByYWlzZSBSdW50aW1lRXJyb3IoZiJ7bGFiZWx9OiBleHBlY3RlZCBleGFjdGx5IG9uZSByZWdleCBtYXRjaCwgZm91bmQge2NvdW50fSIpCiAgICByZXR1cm4gdXBkYXRlZAoKCiMgLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tCiMgUmVncmVzc2lvbnMuCiMgLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tCndyaXRlKAogICAgInRlc3RzL2RhdGEvdGVzdF9mYWxsYmFja190aWNrX3dvcmtlcl9saWZlY3ljbGUucHkiLAogICAgZGVkZW50KCcnJ1wKICAgICAgICBmcm9tIF9fZnV0dXJlX18gaW1wb3J0IGFubm90YXRpb25zCgogICAgICAgIGltcG9ydCB0aHJlYWRpbmcKICAgICAgICBpbXBvcnQgdGltZQoKICAgICAgICBmcm9tIG5pZnR5X3NjYWxwZXJfYm90LmRhdGEubWFya2V0X2RhdGFfbWFuYWdlciBpbXBvcnQgTWFya2V0RGF0YU1hbmFnZXIKCgogICAgICAgIGRlZiB0ZXN0X3N0b3BfdGltZW91dF9wcmVzZXJ2ZXNfb3duZXJfYW5kX2Jsb2Nrc19yZXBsYWNlbWVudCgpIC0+IE5vbmU6CiAgICAgICAgICAgIG1kbSA9IE1hcmtldERhdGFNYW5hZ2VyKGtpdGU9Tm9uZSkKICAgICAgICAgICAgZW50ZXJlZCA9IHRocmVhZGluZy5FdmVudCgpCiAgICAgICAgICAgIHJlbGVhc2UgPSB0aHJlYWRpbmcuRXZlbnQoKQoKICAgICAgICAgICAgZGVmIGJsb2NrZWQoX3RpY2spOgogICAgICAgICAgICAgICAgZW50ZXJlZC5zZXQoKQogICAgICAgICAgICAgICAgcmVsZWFzZS53YWl0KHRpbWVvdXQ9NS4wKQoKICAgICAgICAgICAgbWRtLl9wcm9jZXNzX3F1ZXVlZF90aWNrID0gYmxvY2tlZAogICAgICAgICAgICBtZG0uX3RpY2tfd29ya2VyX3N0b3BfdGltZW91dF9zID0gMC4wNQogICAgICAgICAgICBhc3NlcnQgbWRtLl9lbnN1cmVfdGlja193b3JrZXIoKSBpcyBUcnVlCiAgICAgICAgICAgIG93bmVyID0gbWRtLl90aWNrX3dvcmtlcl90aHJlYWQKICAgICAgICAgICAgbWRtLl9mYWxsYmFja190aWNrX3F1ZXVlLnB1dCh7InN5bWJvbCI6ICJOU0U6TklGVFkifSkKICAgICAgICAgICAgYXNzZXJ0IGVudGVyZWQud2FpdCh0aW1lb3V0PTEuMCkKCiAgICAgICAgICAgIGFzc2VydCBtZG0uX3N0b3BfZmFsbGJhY2tfdGlja193b3JrZXIoKSBpcyBGYWxzZQogICAgICAgICAgICBhc3NlcnQgbWRtLl90aWNrX3dvcmtlcl9zdGF0ZSA9PSAiRkFJTEVEX1RPX1NUT1AiCiAgICAgICAgICAgIGFzc2VydCBtZG0uX3RpY2tfd29ya2VyX3N0b3AuaXNfc2V0KCkKICAgICAgICAgICAgYXNzZXJ0IG1kbS5fdGlja193b3JrZXJfdGhyZWFkIGlzIG93bmVyCiAgICAgICAgICAgIGFzc2VydCBvd25lciBpcyBub3QgTm9uZSBhbmQgb3duZXIuaXNfYWxpdmUoKQogICAgICAgICAgICBhc3NlcnQgbWRtLl9lbnN1cmVfdGlja193b3JrZXIoKSBpcyBGYWxzZQogICAgICAgICAgICBhc3NlcnQgbWRtLl90aWNrX3dvcmtlcl90aHJlYWQgaXMgb3duZXIKCiAgICAgICAgICAgIHJlbGVhc2Uuc2V0KCkKICAgICAgICAgICAgb3duZXIuam9pbih0aW1lb3V0PTEuMCkKICAgICAgICAgICAgZGVhZGxpbmUgPSB0aW1lLm1vbm90b25pYygpICsgMS4wCiAgICAgICAgICAgIHdoaWxlIG1kbS5fdGlja193b3JrZXJfc3RhdGUgIT0gIlNUT1BQRUQiIGFuZCB0aW1lLm1vbm90b25pYygpIDwgZGVhZGxpbmU6CiAgICAgICAgICAgICAgICB0aW1lLnNsZWVwKDAuMDEpCiAgICAgICAgICAgIGFzc2VydCBtZG0uX3RpY2tfd29ya2VyX3N0YXRlID09ICJTVE9QUEVEIgogICAgICAgICAgICBhc3NlcnQgbWRtLl90aWNrX3dvcmtlcl90aHJlYWQgaXMgTm9uZQoKICAgICAgICAgICAgYXNzZXJ0IG1kbS5fZW5zdXJlX3RpY2tfd29ya2VyKCkgaXMgVHJ1ZQogICAgICAgICAgICByZXBsYWNlbWVudCA9IG1kbS5fdGlja193b3JrZXJfdGhyZWFkCiAgICAgICAgICAgIGFzc2VydCByZXBsYWNlbWVudCBpcyBub3QgTm9uZSBhbmQgcmVwbGFjZW1lbnQgaXMgbm90IG93bmVyCiAgICAgICAgICAgIGFzc2VydCBtZG0uX3N0b3BfZmFsbGJhY2tfdGlja193b3JrZXIoKSBpcyBUcnVlCiAgICAnJycpCikKd3JpdGUoCiAgICAidGVzdHMvZXhlY3V0aW9uL3Rlc3RfcG9zaXRpb25fcmVjb25jaWxlX3NpbmdsZWZsaWdodC5weSIsCiAgICBkZWRlbnQoJycnXAogICAgICAgIGZyb20gX19mdXR1cmVfXyBpbXBvcnQgYW5ub3RhdGlvbnMKCiAgICAgICAgaW1wb3J0IHRocmVhZGluZwoKICAgICAgICBpbXBvcnQgcHl0ZXN0CgogICAgICAgIGZyb20gbmlmdHlfc2NhbHBlcl9ib3QuZXhlY3V0aW9uLnBvc2l0aW9uX21hbmFnZXIgaW1wb3J0ICgKICAgICAgICAgICAgUG9zaXRpb25NYW5hZ2VyLAogICAgICAgICAgICBTdGFsZVJlY29uY2lsaWF0aW9uU25hcHNob3QsCiAgICAgICAgKQoKICAgICAgICBTWU1CT0wgPSAiTkZPOk5JRlRZMjZBVUcyNTAwMENFIgogICAgICAgIE9USEVSID0gIk5GTzpOSUZUWTI2QVVHMjUwMDBQRSIKCgogICAgICAgIGNsYXNzIEJsb2NraW5nQnJva2VyOgogICAgICAgICAgICBkZWYgX19pbml0X18oc2VsZiwgcGF5bG9hZCk6CiAgICAgICAgICAgICAgICBzZWxmLnBheWxvYWQgPSBwYXlsb2FkCiAgICAgICAgICAgICAgICBzZWxmLmVudGVyZWQgPSB0aHJlYWRpbmcuRXZlbnQoKQogICAgICAgICAgICAgICAgc2VsZi5yZWxlYXNlID0gdGhyZWFkaW5nLkV2ZW50KCkKICAgICAgICAgICAgICAgIHNlbGYuY2FsbHMgPSAwCgogICAgICAgICAgICBkZWYgZ2V0X3Bvc2l0aW9ucyhzZWxmKToKICAgICAgICAgICAgICAgIHNlbGYuY2FsbHMgKz0gMQogICAgICAgICAgICAgICAgc2VsZi5lbnRlcmVkLnNldCgpCiAgICAgICAgICAgICAgICBzZWxmLnJlbGVhc2Uud2FpdCh0aW1lb3V0PTMuMCkKICAgICAgICAgICAgICAgIHJldHVy
