+++ scripts/p0p1_chunks/tests.part.01
@@ -1,0 +1,1 @@
+biBzZWxmLnBheWxvYWQKCgogICAgICAgIGRlZiB0ZXN0X2NvbmN1cnJlbnRfcmVjb25jaWxlX3JlcXVlc3RzX2FyZV9jb2FsZXNjZWQodG1wX3BhdGgpIC0+IE5vbmU6CiAgICAgICAgICAgIHBtID0gUG9zaXRpb25NYW5hZ2VyKHN0cih0bXBfcGF0aCAvICJwb3NpdGlvbnMuanNvbiIpKQogICAgICAgICAgICBicm9rZXIgPSBCbG9ja2luZ0Jyb2tlcihbXSkKICAgICAgICAgICAgcG0uc2V0X2Jyb2tlcl9jbGllbnQoYnJva2VyKQogICAgICAgICAgICByZXN1bHQ6IGxpc3RbYm9vbF0gPSBbXQogICAgICAgICAgICB3b3JrZXIgPSB0aHJlYWRpbmcuVGhyZWFkKHRhcmdldD1sYW1iZGE6IHJlc3VsdC5hcHBlbmQocG0ucmVjb25jaWxlX25vdygpKSkKICAgICAgICAgICAgd29ya2VyLnN0YXJ0KCkKICAgICAgICAgICAgYXNzZXJ0IGJyb2tlci5lbnRlcmVkLndhaXQodGltZW91dD0xLjApCgogICAgICAgICAgICBhc3NlcnQgcG0ucmVjb25jaWxlX25vdygpIGlzIFRydWUKICAgICAgICAgICAgYXNzZXJ0IGJyb2tlci5jYWxscyA9PSAxCiAgICAgICAgICAgIGFzc2VydCBwbS5fcmVjb25jaWxlX2NvYWxlc2NlZF9yZXF1ZXN0cyA9PSAxCgogICAgICAgICAgICBicm9rZXIucmVsZWFzZS5zZXQoKQogICAgICAgICAgICB3b3JrZXIuam9pbih0aW1lb3V0PTIuMCkKICAgICAgICAgICAgcG0uX2NhbmNlbF9yZWNvbmNpbGVfdGltZXIoKQogICAgICAgICAgICBhc3NlcnQgcmVzdWx0ID09IFtUcnVlXQogICAgICAgICAgICBhc3NlcnQgYnJva2VyLmNhbGxzID09IDEKCgogICAgICAgIGRlZiB0ZXN0X2xvY2FsX2ZpbGxfZHVyaW5nX2ZldGNoX3JlamVjdHNfc3RhbGVfYnJva2VyX3NuYXBzaG90KHRtcF9wYXRoKSAtPiBOb25lOgogICAgICAgICAgICBwbSA9IFBvc2l0aW9uTWFuYWdlcihzdHIodG1wX3BhdGggLyAicG9zaXRpb25zLmpzb24iKSkKICAgICAgICAgICAgcG0ub3Blbl9wb3NpdGlvbihTWU1CT0wsICJMT05HIiwgNjUsIDEwMC4wKQogICAgICAgICAgICBicm9rZXIgPSBCbG9ja2luZ0Jyb2tlcihbXSkKICAgICAgICAgICAgcG0uc2V0X2Jyb2tlcl9jbGllbnQoYnJva2VyKQogICAgICAgICAgICByZXN1bHQ6IGxpc3RbYm9vbF0gPSBbXQogICAgICAgICAgICB3b3JrZXIgPSB0aHJlYWRpbmcuVGhyZWFkKHRhcmdldD1sYW1iZGE6IHJlc3VsdC5hcHBlbmQocG0ucmVjb25jaWxlX25vdygpKSkKICAgICAgICAgICAgd29ya2VyLnN0YXJ0KCkKICAgICAgICAgICAgYXNzZXJ0IGJyb2tlci5lbnRlcmVkLndhaXQodGltZW91dD0xLjApCgogICAgICAgICAgICBwbS5vcGVuX3Bvc2l0aW9uKE9USEVSLCAiTE9ORyIsIDY1LCA5MC4wKQogICAgICAgICAgICBicm9rZXIucmVsZWFzZS5zZXQoKQogICAgICAgICAgICB3b3JrZXIuam9pbih0aW1lb3V0PTIuMCkKICAgICAgICAgICAgcG0uX2NhbmNlbF9yZWNvbmNpbGVfdGltZXIoKQoKICAgICAgICAgICAgYXNzZXJ0IHJlc3VsdCA9PSBbRmFsc2VdCiAgICAgICAgICAgIGFzc2VydCBwbS5nZXRfcG9zaXRpb24oU1lNQk9MKSBpcyBub3QgTm9uZQogICAgICAgICAgICBhc3NlcnQgcG0uZ2V0X3Bvc2l0aW9uKE9USEVSKSBpcyBub3QgTm9uZQogICAgICAgICAgICBhc3NlcnQgcG0uX2xhc3RfcmVjb25jaWxlX2Vycm9yIGlzIG5vdCBOb25lCiAgICAgICAgICAgIGFzc2VydCAiZ2VuZXJhdGlvbiIgaW4gcG0uX2xhc3RfcmVjb25jaWxlX2Vycm9yCgoKICAgICAgICBkZWYgdGVzdF9vbGRlcl9yZWNvbmNpbGVfZ2VuZXJhdGlvbl9jYW5ub3Rfb3ZlcndyaXRlX25ld2VyKHRtcF9wYXRoKSAtPiBOb25lOgogICAgICAgICAgICBwbSA9IFBvc2l0aW9uTWFuYWdlcihzdHIodG1wX3BhdGggLyAicG9zaXRpb25zLmpzb24iKSkKICAgICAgICAgICAgZ2VuZXJhdGlvbiA9IHBtLl9sb2NhbF9wb3NpdGlvbl9nZW5lcmF0aW9uCiAgICAgICAgICAgIHBtLnN5bmNocm9uaXplX3dpdGhfYnJva2VyKAogICAgICAgICAgICAgICAgW10sCiAgICAgICAgICAgICAgICBleHBlY3RlZF9sb2NhbF9nZW5lcmF0aW9uPWdlbmVyYXRpb24sCiAgICAgICAgICAgICAgICByZWNvbmNpbGVfZ2VuZXJhdGlvbj0yLAogICAgICAgICAgICApCiAgICAgICAgICAgIHdpdGggcHl0ZXN0LnJhaXNlcyhTdGFsZVJlY29uY2lsaWF0aW9uU25hcHNob3QpOgogICAgICAgICAgICAgICAgcG0uc3luY2hyb25pemVfd2l0aF9icm9rZXIoCiAgICAgICAgICAgICAgICAgICAgW10sCiAgICAgICAgICAgICAgICAgICAgZXhwZWN0ZWRfbG9jYWxfZ2VuZXJhdGlvbj1nZW5lcmF0aW9uLAogICAgICAgICAgICAgICAgICAgIHJlY29uY2lsZV9nZW5lcmF0aW9uPTEsCiAgICAgICAgICAgICAgICApCiAgICAnJycpCikKd3JpdGUoCiAgICAidGVzdHMvaW5mcmEvdGVzdF9tYXJrZXRfZGF0YV9zdXBlcnZpc29yLnB5IiwKICAgIGRlZGVudCgnJydcCiAgICAgICAgZnJvbSBfX2Z1dHVyZV9fIGltcG9ydCBhbm5vdGF0aW9ucwoKICAgICAgICBmcm9tIHR5cGVzIGltcG9ydCBTaW1wbGVOYW1lc3BhY2UKCiAgICAgICAgZnJvbSBuaWZ0eV9zY2FscGVyX2JvdC5pbmZyYS53YXRjaGRvZyBpbXBvcnQgKAogICAgICAgICAgICBNYXJrZXREYXRhU3VwZXJ2aXNvciwKICAgICAgICAgICAgU3VwZXJ2aXNvclN0YXRlLAogICAgICAgICkKCgogICAgICAgIGNsYXNzIE1ETShTaW1wbGVOYW1lc3BhY2UpOgogICAgICAgICAgICBkZWYgX19pbml0X18oc2VsZik6CiAgICAgICAgICAgICAgICBzdXBlcigpLl9faW5pdF9fKAogICAgICAgICAgICAgICAgICAgIGxhc3RfcG9sbF9oZWFydGJlYXQ9MTAwLjAsCiAgICAgICAgICAgICAgICAgICAgbGFzdF93c19oZWFydGJlYXQ9MTAwLjAsCiAgICAgICAgICAgICAgICAgICAgbGFzdF90aWNrX3RpbWU9MTAwLjAsCiAgICAgICAgICAgICAgICAgICAgX2V2ZW50X2xvb3BfbGFnX21zPTAuMCwKICAgICAgICAgICAgICAgICAgICBfdGlja19xdWV1ZV9tYXhzaXplPTEwMCwKICAgICAgICAgICAgICAgICkKICAgICAgICAgICAgICAgIHNlbGYucG9sbF9yZXN0YXJ0cyA9IDAKICAgICAgICAgICAgICAgIHNlbGYud3NfcmVzdGFydHMgPSAwCiAgICAgICAgICAgICAgICBzZWxmLnByZXNzdXJlID0ge30KCiAgICAgICAgICAgIGRlZiByZXN0YXJ0X3BvbGxpbmcoc2VsZik6CiAgICAgICAgICAgICAgICBzZWxmLnBvbGxfcmVzdGFydHMgKz0gMQoKICAgICAgICAgICAgZGVmIHJlc3RhcnRfd2Vic29ja2V0KHNlbGYpOgogICAgICAgICAgICAgICAgc2VsZi53c19yZXN0YXJ0cyArPSAxCgogICAgICAgICAgICBkZWYgZ2V0X3RpY2tfcHJlc3N1cmVfc3RhdHMoc2VsZik6CiAgICAgICAgICAgICAgICByZXR1cm4gZGljdChzZWxmLnByZXNzdXJlKQoKCiAgICAgICAgZGVmIHRlc3RfcmVzdGFydF9pc19tYXJrZXRfYXdhcmVfYW5kX2JhY2tlZF9vZmYoKSAtPiBOb25lOgogICAgICAgICAgICBtZG0gPSBNRE0oKQogICAgICAgICAgICBzdXBlcnZpc29yID0gTWFya2V0RGF0YVN1cGVydmlzb3IoCiAgICAgICAgICAgICAgICBtZG0sIG1hcmtldF9vcGVuX2dldHRlcj1sYW1iZGE6IFRydWUsIGV4aXRfcHJvY2Vzcz1sYW1iZGEgX2NvZGU6IE5vbmUKICAgICAgICAgICAgKQogICAgICAgICAgICBzdXBlcnZpc29yLnN0YWxlX3Jlc3RhcnRfcyA9IDUuMAogICAgICAgICAgICBzdXBlcnZpc29yLnJlc3RhcnRfYmFzZV9zID0gMTAuMAogICAgICAgICAgICBzdXBlcnZpc29yLnNhbXBsZV9vbmNlKG5vd193YWxsPTExMC4wLCBub3dfbW9ubz01MC4wKQogICAgICAgICAgICBmaXJzdCA9
