+++ scripts/p0p1_chunks/tests.part.02
@@ -1,0 +1,1 @@
+IG1kbS5wb2xsX3Jlc3RhcnRzICsgbWRtLndzX3Jlc3RhcnRzCiAgICAgICAgICAgIGFzc2VydCBmaXJzdCA+PSAxCiAgICAgICAgICAgIHN1cGVydmlzb3Iuc2FtcGxlX29uY2Uobm93X3dhbGw9MTExLjAsIG5vd19tb25vPTUxLjApCiAgICAgICAgICAgIGFzc2VydCBtZG0ucG9sbF9yZXN0YXJ0cyArIG1kbS53c19yZXN0YXJ0cyA9PSBmaXJzdAoKICAgICAgICAgICAgY2xvc2VkID0gTWFya2V0RGF0YVN1cGVydmlzb3IoCiAgICAgICAgICAgICAgICBtZG0sIG1hcmtldF9vcGVuX2dldHRlcj1sYW1iZGE6IEZhbHNlLCBleGl0X3Byb2Nlc3M9bGFtYmRhIF9jb2RlOiBOb25lCiAgICAgICAgICAgICkKICAgICAgICAgICAgYmVmb3JlID0gbWRtLnBvbGxfcmVzdGFydHMgKyBtZG0ud3NfcmVzdGFydHMKICAgICAgICAgICAgc25hcHNob3QgPSBjbG9zZWQuc2FtcGxlX29uY2Uobm93X3dhbGw9MjAwLjAsIG5vd19tb25vPTIwMC4wKQogICAgICAgICAgICBhc3NlcnQgc25hcHNob3Quc3RhdGUgPT0gU3VwZXJ2aXNvclN0YXRlLk1BUktFVF9DTE9TRUQudmFsdWUKICAgICAgICAgICAgYXNzZXJ0IG1kbS5wb2xsX3Jlc3RhcnRzICsgbWRtLndzX3Jlc3RhcnRzID09IGJlZm9yZQoKCiAgICAgICAgZGVmIHRlc3RfbGF0ZW5jeV9yZXF1aXJlc19jb25zZWN1dGl2ZV9zYW1wbGVzX2FuZF9yZWNvdmVycygpIC0+IE5vbmU6CiAgICAgICAgICAgIG1kbSA9IE1ETSgpCiAgICAgICAgICAgIHN1cGVydmlzb3IgPSBNYXJrZXREYXRhU3VwZXJ2aXNvcigKICAgICAgICAgICAgICAgIG1kbSwgbWFya2V0X29wZW5fZ2V0dGVyPWxhbWJkYTogVHJ1ZSwgZXhpdF9wcm9jZXNzPWxhbWJkYSBfY29kZTogTm9uZQogICAgICAgICAgICApCiAgICAgICAgICAgIHN1cGVydmlzb3IubGF0ZW5jeV9jcml0aWNhbF9zYW1wbGVzID0gMgogICAgICAgICAgICBzdXBlcnZpc29yLmxhdGVuY3lfcmVjb3Zlcnlfc2FtcGxlcyA9IDIKICAgICAgICAgICAgc3VwZXJ2aXNvci5ldmVudF9sb29wX2NyaXRpY2FsX21zID0gMTAwLjAKICAgICAgICAgICAgbWRtLl9ldmVudF9sb29wX2xhZ19tcyA9IDE1MC4wCiAgICAgICAgICAgIHN1cGVydmlzb3Iuc2FtcGxlX29uY2Uobm93X3dhbGw9MTAxLjAsIG5vd19tb25vPTEuMCkKICAgICAgICAgICAgYXNzZXJ0IG1kbS5sYXRlbmN5X2RlZ3JhZGVkIGlzIEZhbHNlCiAgICAgICAgICAgIHN1cGVydmlzb3Iuc2FtcGxlX29uY2Uobm93X3dhbGw9MTAyLjAsIG5vd19tb25vPTIuMCkKICAgICAgICAgICAgYXNzZXJ0IG1kbS5sYXRlbmN5X2RlZ3JhZGVkIGlzIFRydWUKCiAgICAgICAgICAgIG1kbS5fZXZlbnRfbG9vcF9sYWdfbXMgPSAwLjAKICAgICAgICAgICAgc3VwZXJ2aXNvci5zYW1wbGVfb25jZShub3dfd2FsbD0xMDMuMCwgbm93X21vbm89My4wKQogICAgICAgICAgICBhc3NlcnQgbWRtLmxhdGVuY3lfZGVncmFkZWQgaXMgVHJ1ZQogICAgICAgICAgICBzdXBlcnZpc29yLnNhbXBsZV9vbmNlKG5vd193YWxsPTEwNC4wLCBub3dfbW9ubz00LjApCiAgICAgICAgICAgIGFzc2VydCBtZG0ubGF0ZW5jeV9kZWdyYWRlZCBpcyBGYWxzZQoKCiAgICAgICAgZGVmIHRlc3RfZmF0YWxfc3RhbGVfcmVxdWlyZXNfY29uZmlybWF0aW9uKCkgLT4gTm9uZToKICAgICAgICAgICAgbWRtID0gTURNKCkKICAgICAgICAgICAgZXhpdHM6IGxpc3RbaW50XSA9IFtdCiAgICAgICAgICAgIHN1cGVydmlzb3IgPSBNYXJrZXREYXRhU3VwZXJ2aXNvcigKICAgICAgICAgICAgICAgIG1kbSwgbWFya2V0X29wZW5fZ2V0dGVyPWxhbWJkYTogVHJ1ZSwgZXhpdF9wcm9jZXNzPWV4aXRzLmFwcGVuZAogICAgICAgICAgICApCiAgICAgICAgICAgIHN1cGVydmlzb3Iuc3RhbGVfcmVzdGFydF9zID0gMTAwMC4wCiAgICAgICAgICAgIHN1cGVydmlzb3IuZmF0YWxfc3RhbGVfcyA9IDEwLjAKICAgICAgICAgICAgc3VwZXJ2aXNvci5mYXRhbF9jb25maXJtX3NhbXBsZXMgPSAyCiAgICAgICAgICAgIHN1cGVydmlzb3Iuc2FtcGxlX29uY2Uobm93X3dhbGw9MTIwLjAsIG5vd19tb25vPTEuMCkKICAgICAgICAgICAgYXNzZXJ0IGV4aXRzID09IFtdCiAgICAgICAgICAgIHN1cGVydmlzb3Iuc2FtcGxlX29uY2Uobm93X3dhbGw9MTIxLjAsIG5vd19tb25vPTIuMCkKICAgICAgICAgICAgYXNzZXJ0IGV4aXRzID09IFsxXQogICAgICAgICAgICBhc3NlcnQgc3VwZXJ2aXNvci5zdGF0ZSBpcyBTdXBlcnZpc29yU3RhdGUuRkFJTEVECiAgICAnJycpCikKCnBhdGggPSAidGVzdHMvY29yZS90ZXN0X21pbmltdW1fbG90X2FmZm9yZGFiaWxpdHlfcmVhZGluZXNzLnB5Igp0ZXh0ID0gcmVhZChwYXRoKQp0ZXh0ICs9IGRlZGVudCgnJydcCgogICAgQHB5dGVzdC5tYXJrLmFzeW5jaW8KICAgIGFzeW5jIGRlZiB0ZXN0X2NyaXRpY2FsX21hcmtldF9kYXRhX2xhdGVuY3lfZGlzYXJtc19saXZlX29yZGVycyhtb25rZXlwYXRjaCkgLT4gTm9uZToKICAgICAgICBtb25rZXlwYXRjaC5zZXRhdHRyKGFwcCwgImdldF9tYXJrZXRfc3RhdGUiLCBsYW1iZGE6IGFwcC5NYXJrZXRTdGF0ZS5PUEVOKQogICAgICAgIGN0eCwgY2FsbHMgPSBfY29udGV4dCh7Q0U6IDEwMC4wLCBQRTogMTAwLjB9LCAyMF8wMDAuMCkKICAgICAgICBjdHgubWFya2V0X2RhdGFfbWFuYWdlci5sYXRlbmN5X2RlZ3JhZGVkID0gVHJ1ZQoKICAgICAgICBhd2FpdCBhcHAuX3JlY29tcHV0ZV9hbmRfcHVzaF9ydW50aW1lX3JlYWRpbmVzcyhjdHgsIHJlYXNvbj0ibGF0ZW5jeV9zbG9fdGVzdCIpCgogICAgICAgIGFzc2VydCBjdHguZXhlY3V0aW9uX2NhcGFjaXR5X3JlYWR5IGlzIFRydWUKICAgICAgICBhc3NlcnQgY3R4Lm1hcmtldF9kYXRhX2xhdGVuY3lfZGVncmFkZWQgaXMgVHJ1ZQogICAgICAgIGFzc2VydCBjdHgubGl2ZV9vcmRlcnNfYXJtZWQgaXMgRmFsc2UKICAgICAgICBhc3NlcnQgY3R4LmxpdmVfYmxvY2tfcmVhc29uID09ICJleGVjdXRpb25fbm90X2FybWVkOm1hcmtldF9kYXRhX2xhdGVuY3lfZGVncmFkZWQiCiAgICAgICAgYXNzZXJ0IGNhbGxzWy0xXVsibGl2ZV9vcmRlcnNfYXJtZWQiXSBpcyBGYWxzZQonJycpCndyaXRlKHBhdGgsIHRleHQpCgpwcmludCgiUDAvUDEgcnVudGltZSBoYXJkZW5pbmcgYXBwbGllZCIpCg==
