+++ scripts/p0p1_chunks/watchdog.part.00
@@ -1,0 +1,1 @@
+ZnJvbSBfX2Z1dHVyZV9fIGltcG9ydCBhbm5vdGF0aW9ucwoKaW1wb3J0IHJlCmZyb20gcGF0aGxpYiBpbXBvcnQgUGF0aApmcm9tIHRleHR3cmFwIGltcG9ydCBkZWRlbnQKClJPT1QgPSBQYXRoLmN3ZCgpCgoKZGVmIHJlYWQocGF0aDogc3RyKSAtPiBzdHI6CiAgICByZXR1cm4gKFJPT1QgLyBwYXRoKS5yZWFkX3RleHQoZW5jb2Rpbmc9InV0Zi04IikKCgpkZWYgd3JpdGUocGF0aDogc3RyLCB0ZXh0OiBzdHIpIC0+IE5vbmU6CiAgICB0YXJnZXQgPSBST09UIC8gcGF0aAogICAgdGFyZ2V0LnBhcmVudC5ta2RpcihwYXJlbnRzPVRydWUsIGV4aXN0X29rPVRydWUpCiAgICB0YXJnZXQud3JpdGVfdGV4dCh0ZXh0LCBlbmNvZGluZz0idXRmLTgiKQoKCmRlZiByZXBsYWNlX29uY2UodGV4dDogc3RyLCBvbGQ6IHN0ciwgbmV3OiBzdHIsICosIGxhYmVsOiBzdHIpIC0+IHN0cjoKICAgIGNvdW50ID0gdGV4dC5jb3VudChvbGQpCiAgICBpZiBjb3VudCAhPSAxOgogICAgICAgIHJhaXNlIFJ1bnRpbWVFcnJvcihmIntsYWJlbH06IGV4cGVjdGVkIGV4YWN0bHkgb25lIG1hdGNoLCBmb3VuZCB7Y291bnR9IikKICAgIHJldHVybiB0ZXh0LnJlcGxhY2Uob2xkLCBuZXcsIDEpCgoKZGVmIHJlZ2V4X29uY2UodGV4dDogc3RyLCBwYXR0ZXJuOiBzdHIsIHJlcGxhY2VtZW50OiBzdHIsICosIGxhYmVsOiBzdHIsIGZsYWdzOiBpbnQgPSAwKSAtPiBzdHI6CiAgICB1cGRhdGVkLCBjb3VudCA9IHJlLnN1Ym4ocGF0dGVybiwgcmVwbGFjZW1lbnQsIHRleHQsIGNvdW50PTEsIGZsYWdzPWZsYWdzKQogICAgaWYgY291bnQgIT0gMToKICAgICAgICByYWlzZSBSdW50aW1lRXJyb3IoZiJ7bGFiZWx9OiBleHBlY3RlZCBleGFjdGx5IG9uZSByZWdleCBtYXRjaCwgZm91bmQge2NvdW50fSIpCiAgICByZXR1cm4gdXBkYXRlZAoKCiMgLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tCiMgUDE6IG9uZSBtYXJrZXQtYXdhcmUgc3VwZXJ2aXNvciB3aXRoIHJlc3RhcnQgYmFja29mZiBhbmQgbGF0ZW5jeSBTTE8gc3RhdGUuCiMgLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tCndhdGNoZG9nID0gZGVkZW50KCcnJ1wKICAgICIiIlNpbmdsZS1vd25lciBtYXJrZXQtZGF0YSBzdXBlcnZpc2lvbiBhbmQgbGF0ZW5jeSBTTE8gZW5mb3JjZW1lbnQuIiIiCgogICAgZnJvbSBfX2Z1dHVyZV9fIGltcG9ydCBhbm5vdGF0aW9ucwoKICAgIGZyb20gZGF0YWNsYXNzZXMgaW1wb3J0IGFzZGljdCwgZGF0YWNsYXNzCiAgICBmcm9tIGVudW0gaW1wb3J0IEVudW0KICAgIGltcG9ydCBsb2dnaW5nCiAgICBpbXBvcnQgb3MKICAgIGltcG9ydCB0aHJlYWRpbmcKICAgIGltcG9ydCB0aW1lCiAgICBmcm9tIHR5cGluZyBpbXBvcnQgQW55LCBDYWxsYWJsZSwgTWFwcGluZwoKICAgIExPR0dFUiA9IGxvZ2dpbmcuZ2V0TG9nZ2VyKCJuaWZ0eV9zY2FscGVyX2JvdC5tYXJrZXRfZGF0YV9zdXBlcnZpc29yIikKCgogICAgY2xhc3MgU3VwZXJ2aXNvclN0YXRlKHN0ciwgRW51bSk6CiAgICAgICAgU1RBUlRJTkcgPSAiU1RBUlRJTkciCiAgICAgICAgSEVBTFRIWSA9ICJIRUFMVEhZIgogICAgICAgIERFR1JBREVEID0gIkRFR1JBREVEIgogICAgICAgIFJFQ09OTkVDVElORyA9ICJSRUNPTk5FQ1RJTkciCiAgICAgICAgTUFSS0VUX0NMT1NFRCA9ICJNQVJLRVRfQ0xPU0VEIgogICAgICAgIEZBSUxFRCA9ICJGQUlMRUQiCiAgICAgICAgU1RPUFBJTkcgPSAiU1RPUFBJTkciCiAgICAgICAgU1RPUFBFRCA9ICJTVE9QUEVEIgoKCiAgICBAZGF0YWNsYXNzKGZyb3plbj1UcnVlLCBzbG90cz1UcnVlKQogICAgY2xhc3MgU3VwZXJ2aXNvclNuYXBzaG90OgogICAgICAgIHN0YXRlOiBzdHIKICAgICAgICBsYXRlbmN5X2RlZ3JhZGVkOiBib29sCiAgICAgICAgZXZlbnRfbG9vcF9sYWdfbXM6IGZsb2F0CiAgICAgICAgdGlja19wOTlfbXM6IGZsb2F0CiAgICAgICAgcXVldWVfYWdlX21zOiBmbG9hdAogICAgICAgIHF1ZXVlX3V0aWxpemF0aW9uOiBmbG9hdAogICAgICAgIG9wZW5fcG9zaXRpb25fcXVvdGVfYWdlX3M6IGZsb2F0CiAgICAgICAgY29uc2VjdXRpdmVfZmFpbHVyZXM6IGludAogICAgICAgIHJlc3RhcnRfY291bnQ6IGludAogICAgICAgIGZhdGFsX3N0YWxlX3NhbXBsZXM6IGludAoKICAgICAgICBkZWYgdG9fZGljdChzZWxmKSAtPiBkaWN0W3N0ciwgb2JqZWN0XToKICAgICAgICAgICAgcmV0dXJuIGFzZGljdChzZWxmKQoKCiAgICBkZWYgX2Zsb2F0X2VudihuYW1lOiBzdHIsIGRlZmF1bHQ6IGZsb2F0LCBtaW5pbXVtOiBmbG9hdCA9IDAuMCkgLT4gZmxvYXQ6CiAgICAgICAgdHJ5OgogICAgICAgICAgICB2YWx1ZSA9IGZsb2F0KG9zLmdldGVudihuYW1lLCBzdHIoZGVmYXVsdCkpIG9yIGRlZmF1bHQpCiAgICAgICAgZXhjZXB0IChUeXBlRXJyb3IsIFZhbHVlRXJyb3IpOgogICAgICAgICAgICB2YWx1ZSA9IGRlZmF1bHQKICAgICAgICByZXR1cm4gbWF4KHZhbHVlLCBtaW5pbXVtKQoKCiAgICBkZWYgX2ludF9lbnYobmFtZTogc3RyLCBkZWZhdWx0OiBpbnQsIG1pbmltdW06IGludCA9IDEpIC0+IGludDoKICAgICAgICB0cnk6CiAgICAgICAgICAgIHZhbHVlID0gaW50KG9zLmdldGVudihuYW1lLCBzdHIoZGVmYXVsdCkpIG9yIGRlZmF1bHQpCiAgICAgICAgZXhjZXB0IChUeXBlRXJyb3IsIFZhbHVlRXJyb3IpOgogICAgICAgICAgICB2YWx1ZSA9IGRlZmF1bHQKICAgICAgICByZXR1cm4gbWF4KHZhbHVlLCBtaW5pbXVtKQoKCiAgICBkZWYgX2RlZmF1bHRfbWFya2V0X29wZW4oKSAtPiBib29sOgogICAgICAgIHRyeToKICAgICAgICAgICAgZnJvbSBuaWZ0eV9zY2FscGVyX2JvdC51dGlscy5tYXJrZXRfaG91cnMgaW1wb3J0IGdldF9ydW50aW1lX21hcmtldF9tb2RlCgogICAgICAgICAgICB0b2tlbiA9IHN0cihnZXRfcnVudGltZV9tYXJrZXRfbW9kZSgpKS51cHBlcigpCiAgICAgICAgICAgIHJldHVybiAiT1BFTiIgaW4gdG9rZW4gYW5kICJQT1NUIiBub3QgaW4gdG9rZW4KICAgICAgICBleGNlcHQgRXhjZXB0aW9uOgogICAgICAgICAgICByZXR1cm4gVHJ1ZQoKCiAgICBkZWYgX2FnZV9zZWNvbmRzKHJhdzogb2JqZWN0LCBub3c6IGZsb2F0KSAtPiBmbG9hdCB8IE5vbmU6CiAgICAgICAgdHJ5OgogICAgICAgICAgICB2YWx1ZSA9IGZsb2F0KHJhdyBvciAwLjApCiAgICAgICAgZXhjZXB0IChUeXBlRXJyb3IsIFZhbHVlRXJyb3IpOgogICAgICAgICAgICByZXR1cm4gTm9uZQogICAgICAgIGlmIHZhbHVlIDw9IDA6CiAgICAgICAgICAgIHJldHVybiBOb25lCiAgICAgICAgcmV0dXJuIG1heCgwLjAsIG5vdyAtIHZhbHVlKQoKCiAgICBjbGFzcyBNYXJrZXREYXRhU3VwZXJ2aXNvcjoKICAgICAgICAiIiJPbmUgZGV0ZXJtaW5pc3RpYyBvd25lciBmb3IgcmVzdGFydCBwb2xpY3kgYW5kIGxhdGVuY3kgZGVncmFkYXRpb24uIiIiCgogICAgICAgIGRlZiBfX2luaXRfXygKICAgICAgICAgICAgc2VsZiwKICAgICAgICAgICAgbWFya2V0X2RhdGFfbWFuYWdlcjogQW55LAogICAgICAgICAgICAqLAogICAgICAgICAgICBtYXJrZXRfb3Blbl9nZXR0ZXI6IENhbGxhYmxlW1tdLCBib29sXSA9IF9kZWZhdWx0X21hcmtldF9vcGVuLAogICAgICAgICAgICBleGl0X3Byb2Nlc3M6IENhbGxhYmxlW1tpbnRdLCBBbnldID0gb3MuX2V4aXQsCiAgICAgICAgICAgIGludGVy
