+++ scripts/p0p1_chunks/watchdog.part.01
@@ -1,0 +1,1 @@
+dmFsX3M6IGZsb2F0IHwgTm9uZSA9IE5vbmUsCiAgICAgICAgKSAtPiBOb25lOgogICAgICAgICAgICBzZWxmLm1kbSA9IG1hcmtldF9kYXRhX21hbmFnZXIKICAgICAgICAgICAgc2VsZi5tYXJrZXRfb3Blbl9nZXR0ZXIgPSBtYXJrZXRfb3Blbl9nZXR0ZXIKICAgICAgICAgICAgc2VsZi5leGl0X3Byb2Nlc3MgPSBleGl0X3Byb2Nlc3MKICAgICAgICAgICAgc2VsZi5pbnRlcnZhbF9zID0gaW50ZXJ2YWxfcyBvciBfZmxvYXRfZW52KAogICAgICAgICAgICAgICAgIk1BUktFVF9EQVRBX1NVUEVSVklTT1JfSU5URVJWQUxfU0VDT05EUyIsIDEuMCwgMC4xCiAgICAgICAgICAgICkKICAgICAgICAgICAgc2VsZi5zdGFsZV9yZXN0YXJ0X3MgPSBfZmxvYXRfZW52KAogICAgICAgICAgICAgICAgIk1BUktFVF9EQVRBX1JFU1RBUlRfU1RBTEVfU0VDT05EUyIsIDE1LjAsIDEuMAogICAgICAgICAgICApCiAgICAgICAgICAgIHNlbGYuZmF0YWxfc3RhbGVfcyA9IF9mbG9hdF9lbnYoCiAgICAgICAgICAgICAgICAiTUFSS0VUX0RBVEFfRkFUQUxfU1RBTEVfU0VDT05EUyIsIDE4MC4wLCAxMC4wCiAgICAgICAgICAgICkKICAgICAgICAgICAgc2VsZi5yZXN0YXJ0X2Jhc2VfcyA9IF9mbG9hdF9lbnYoCiAgICAgICAgICAgICAgICAiTUFSS0VUX0RBVEFfUkVTVEFSVF9CQUNLT0ZGX1NFQ09ORFMiLCA1LjAsIDAuMQogICAgICAgICAgICApCiAgICAgICAgICAgIHNlbGYucmVzdGFydF9tYXhfcyA9IF9mbG9hdF9lbnYoCiAgICAgICAgICAgICAgICAiTUFSS0VUX0RBVEFfUkVTVEFSVF9CQUNLT0ZGX01BWF9TRUNPTkRTIiwgMTIwLjAsIDEuMAogICAgICAgICAgICApCiAgICAgICAgICAgIHNlbGYuZmF0YWxfY29uZmlybV9zYW1wbGVzID0gX2ludF9lbnYoCiAgICAgICAgICAgICAgICAiTUFSS0VUX0RBVEFfRkFUQUxfQ09ORklSTV9TQU1QTEVTIiwgMwogICAgICAgICAgICApCiAgICAgICAgICAgIHNlbGYubGF0ZW5jeV9jcml0aWNhbF9zYW1wbGVzID0gX2ludF9lbnYoCiAgICAgICAgICAgICAgICAiTUFSS0VUX0RBVEFfTEFURU5DWV9DUklUSUNBTF9TQU1QTEVTIiwgMwogICAgICAgICAgICApCiAgICAgICAgICAgIHNlbGYubGF0ZW5jeV9yZWNvdmVyeV9zYW1wbGVzID0gX2ludF9lbnYoCiAgICAgICAgICAgICAgICAiTUFSS0VUX0RBVEFfTEFURU5DWV9SRUNPVkVSWV9TQU1QTEVTIiwgNQogICAgICAgICAgICApCiAgICAgICAgICAgIHNlbGYuZXZlbnRfbG9vcF9jcml0aWNhbF9tcyA9IF9mbG9hdF9lbnYoCiAgICAgICAgICAgICAgICAiRVZFTlRfTE9PUF9MQUdfQ1JJVElDQUxfTVMiLCA1MDAuMCwgMS4wCiAgICAgICAgICAgICkKICAgICAgICAgICAgc2VsZi50aWNrX3A5OV9jcml0aWNhbF9tcyA9IF9mbG9hdF9lbnYoCiAgICAgICAgICAgICAgICAiVElDS19QOTlfQ1JJVElDQUxfTVMiLCAyNTAuMCwgMS4wCiAgICAgICAgICAgICkKICAgICAgICAgICAgc2VsZi5xdWV1ZV9hZ2VfY3JpdGljYWxfbXMgPSBfZmxvYXRfZW52KAogICAgICAgICAgICAgICAgIlRJQ0tfUVVFVUVfQUdFX0NSSVRJQ0FMX01TIiwgMTAwMC4wLCAxLjAKICAgICAgICAgICAgKQogICAgICAgICAgICBzZWxmLnF1ZXVlX3V0aWxpemF0aW9uX2NyaXRpY2FsID0gbWluKAogICAgICAgICAgICAgICAgX2Zsb2F0X2VudigiVElDS19RVUVVRV9VVElMSVpBVElPTl9DUklUSUNBTCIsIDAuODUsIDAuMDEpLCAxLjAKICAgICAgICAgICAgKQogICAgICAgICAgICBzZWxmLm9wZW5fcXVvdGVfYWdlX2NyaXRpY2FsX3MgPSBfZmxvYXRfZW52KAogICAgICAgICAgICAgICAgIk9QRU5fUE9TSVRJT05fUVVPVEVfQUdFX0NSSVRJQ0FMX1NFQ09ORFMiLCAyLjAsIDAuMQogICAgICAgICAgICApCiAgICAgICAgICAgIHNlbGYuX3N0b3AgPSB0aHJlYWRpbmcuRXZlbnQoKQogICAgICAgICAgICBzZWxmLl90aHJlYWQ6IHRocmVhZGluZy5UaHJlYWQgfCBOb25lID0gTm9uZQogICAgICAgICAgICBzZWxmLl9sb2NrID0gdGhyZWFkaW5nLlJMb2NrKCkKICAgICAgICAgICAgc2VsZi5fcmVzdGFydF9sb2NrID0gdGhyZWFkaW5nLkxvY2soKQogICAgICAgICAgICBzZWxmLl9zdGF0ZSA9IFN1cGVydmlzb3JTdGF0ZS5TVEFSVElORwogICAgICAgICAgICBzZWxmLl9jb25zZWN1dGl2ZV9mYWlsdXJlcyA9IDAKICAgICAgICAgICAgc2VsZi5fcmVzdGFydF9jb3VudCA9IDAKICAgICAgICAgICAgc2VsZi5fbmV4dF9yZXN0YXJ0X21vbm8gPSAwLjAKICAgICAgICAgICAgc2VsZi5fZmF0YWxfc3RhbGVfc2FtcGxlcyA9IDAKICAgICAgICAgICAgc2VsZi5fY3JpdGljYWxfbGF0ZW5jeV9zYW1wbGVzID0gMAogICAgICAgICAgICBzZWxmLl9oZWFsdGh5X2xhdGVuY3lfc2FtcGxlcyA9IDAKICAgICAgICAgICAgc2VsZi5fbGF0ZW5jeV9kZWdyYWRlZCA9IEZhbHNlCiAgICAgICAgICAgIHNlbGYuX3NuYXBzaG90OiBTdXBlcnZpc29yU25hcHNob3QgfCBOb25lID0gTm9uZQoKICAgICAgICBAcHJvcGVydHkKICAgICAgICBkZWYgdGhyZWFkKHNlbGYpIC0+IHRocmVhZGluZy5UaHJlYWQgfCBOb25lOgogICAgICAgICAgICByZXR1cm4gc2VsZi5fdGhyZWFkCgogICAgICAgIEBwcm9wZXJ0eQogICAgICAgIGRlZiBzdGF0ZShzZWxmKSAtPiBTdXBlcnZpc29yU3RhdGU6CiAgICAgICAgICAgIHdpdGggc2VsZi5fbG9jazoKICAgICAgICAgICAgICAgIHJldHVybiBzZWxmLl9zdGF0ZQoKICAgICAgICBkZWYgc3RhcnQoc2VsZikgLT4gdGhyZWFkaW5nLlRocmVhZDoKICAgICAgICAgICAgd2l0aCBzZWxmLl9sb2NrOgogICAgICAgICAgICAgICAgaWYgc2VsZi5fdGhyZWFkIGlzIG5vdCBOb25lIGFuZCBzZWxmLl90aHJlYWQuaXNfYWxpdmUoKToKICAgICAgICAgICAgICAgICAgICByZXR1cm4gc2VsZi5fdGhyZWFkCiAgICAgICAgICAgICAgICBzZWxmLl9zdG9wLmNsZWFyKCkKICAgICAgICAgICAgICAgIHNlbGYuX3N0YXRlID0gU3VwZXJ2aXNvclN0YXRlLlNUQVJUSU5HCiAgICAgICAgICAgICAgICBzZWxmLl90aHJlYWQgPSB0aHJlYWRpbmcuVGhyZWFkKAogICAgICAgICAgICAgICAgICAgIHRhcmdldD1zZWxmLl9ydW4sCiAgICAgICAgICAgICAgICAgICAgbmFtZT0ibWFya2V0LWRhdGEtc3VwZXJ2aXNvciIsCiAgICAgICAgICAgICAgICAgICAgZGFlbW9uPVRydWUsCiAgICAgICAgICAgICAgICApCiAgICAgICAgICAgICAgICBzZWxmLl90aHJlYWQuc3RhcnQoKQogICAgICAgICAgICAgICAgcmV0dXJuIHNlbGYuX3RocmVhZAoKICAgICAgICBkZWYgc3RvcChzZWxmLCB0aW1lb3V0OiBmbG9hdCA9IDMuMCkgLT4gYm9vbDoKICAgICAgICAgICAgd2l0aCBzZWxmLl9sb2NrOgogICAgICAgICAgICAgICAgc2VsZi5fc3RhdGUgPSBTdXBlcnZpc29yU3RhdGUuU1RPUFBJTkcKICAgICAgICAgICAgICAgIHRocmVhZCA9IHNlbGYuX3RocmVhZAogICAgICAgICAgICAgICAgc2VsZi5fc3RvcC5zZXQoKQogICAgICAgICAgICBpZiB0aHJlYWQgaXMgbm90IE5vbmUgYW5kIHRocmVhZCBpcyBub3QgdGhyZWFkaW5nLmN1cnJlbnRfdGhyZWFkKCk6CiAgICAgICAgICAgICAgICB0aHJlYWQuam9pbih0aW1lb3V0PW1heCh0aW1lb3V0LCAwLjA1KSkKICAgICAgICAgICAgYWxpdmUgPSBib29sKHRocmVhZCBpcyBub3QgTm9uZSBhbmQgdGhyZWFkLmlzX2FsaXZlKCkpCiAgICAgICAgICAgIHdpdGggc2VsZi5fbG9jazoKICAgICAgICAgICAgICAgIGlmIG5vdCBhbGl2ZToKICAgICAgICAgICAgICAgICAgICBzZWxmLl90aHJlYWQgPSBOb25lCiAgICAgICAgICAg
