+++ scripts/p0p1_chunks/watchdog.part.02
@@ -1,0 +1,1 @@
+ICAgICAgICAgc2VsZi5fc3RhdGUgPSBTdXBlcnZpc29yU3RhdGUuU1RPUFBFRAogICAgICAgICAgICByZXR1cm4gbm90IGFsaXZlCgogICAgICAgIGRlZiBfdHJhbnNpdGlvbihzZWxmLCBzdGF0ZTogU3VwZXJ2aXNvclN0YXRlLCByZWFzb246IHN0cikgLT4gTm9uZToKICAgICAgICAgICAgd2l0aCBzZWxmLl9sb2NrOgogICAgICAgICAgICAgICAgcHJldmlvdXMgPSBzZWxmLl9zdGF0ZQogICAgICAgICAgICAgICAgc2VsZi5fc3RhdGUgPSBzdGF0ZQogICAgICAgICAgICBpZiBwcmV2aW91cyAhPSBzdGF0ZToKICAgICAgICAgICAgICAgIExPR0dFUi5pbmZvKAogICAgICAgICAgICAgICAgICAgICJNQVJLRVRfREFUQV9TVVBFUlZJU09SX1NUQVRFIHByZXZpb3VzPSVzIHN0YXRlPSVzIHJlYXNvbj0lcyIsCiAgICAgICAgICAgICAgICAgICAgcHJldmlvdXMudmFsdWUsCiAgICAgICAgICAgICAgICAgICAgc3RhdGUudmFsdWUsCiAgICAgICAgICAgICAgICAgICAgcmVhc29uLAogICAgICAgICAgICAgICAgICAgIGV4dHJhPXsKICAgICAgICAgICAgICAgICAgICAgICAgImV2ZW50IjogIk1BUktFVF9EQVRBX1NVUEVSVklTT1JfU1RBVEUiLAogICAgICAgICAgICAgICAgICAgICAgICAicHJldmlvdXMiOiBwcmV2aW91cy52YWx1ZSwKICAgICAgICAgICAgICAgICAgICAgICAgInN0YXRlIjogc3RhdGUudmFsdWUsCiAgICAgICAgICAgICAgICAgICAgICAgICJyZWFzb24iOiByZWFzb24sCiAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICkKCiAgICAgICAgZGVmIF9wcmVzc3VyZShzZWxmKSAtPiBNYXBwaW5nW3N0ciwgQW55XToKICAgICAgICAgICAgZ2V0dGVyID0gZ2V0YXR0cihzZWxmLm1kbSwgImdldF90aWNrX3ByZXNzdXJlX3N0YXRzIiwgTm9uZSkKICAgICAgICAgICAgaWYgY2FsbGFibGUoZ2V0dGVyKToKICAgICAgICAgICAgICAgIHRyeToKICAgICAgICAgICAgICAgICAgICB2YWx1ZSA9IGdldHRlcigpIG9yIHt9CiAgICAgICAgICAgICAgICAgICAgaWYgaXNpbnN0YW5jZSh2YWx1ZSwgTWFwcGluZyk6CiAgICAgICAgICAgICAgICAgICAgICAgIHJldHVybiB2YWx1ZQogICAgICAgICAgICAgICAgZXhjZXB0IEV4Y2VwdGlvbjoKICAgICAgICAgICAgICAgICAgICBwYXNzCiAgICAgICAgICAgIHJldHVybiB7fQoKICAgICAgICBkZWYgX2xhdGVuY3lfdmFsdWVzKHNlbGYpIC0+IHR1cGxlW2Zsb2F0LCBmbG9hdCwgZmxvYXQsIGZsb2F0LCBmbG9hdF06CiAgICAgICAgICAgIHByZXNzdXJlID0gc2VsZi5fcHJlc3N1cmUoKQogICAgICAgICAgICBldmVudF9sb29wID0gZmxvYXQoZ2V0YXR0cihzZWxmLm1kbSwgIl9ldmVudF9sb29wX2xhZ19tcyIsIDAuMCkgb3IgMC4wKQogICAgICAgICAgICB0aWNrX3A5OSA9IGZsb2F0KAogICAgICAgICAgICAgICAgcHJlc3N1cmUuZ2V0KCJ0aWNrX3A5OV9tcyIpCiAgICAgICAgICAgICAgICBvciBwcmVzc3N1cmUuZ2V0KCJvbmVfdGlja19wOTlfbXMiKQogICAgICAgICAgICAgICAgb3IgZ2V0YXR0cihzZWxmLm1kbSwgInRpY2tfcDk5X21zIiwgMC4wKQogICAgICAgICAgICAgICAgb3IgMC4wCiAgICAgICAgICAgICkKICAgICAgICAgICAgcXVldWVfYWdlID0gZmxvYXQoCiAgICAgICAgICAgICAgICBwcmVzc3N1cmUuZ2V0KCJvbGRlc3RfcGVuZGluZ190aWNrX2FnZV9tcyIpCiAgICAgICAgICAgICAgICBvciBwcmVzc3N1cmUuZ2V0KCJxdWV1ZV9hZ2VfbXMiKQogICAgICAgICAgICAgICAgb3IgMC4wCiAgICAgICAgICAgICkKICAgICAgICAgICAgZGVwdGggPSBmbG9hdChwcmVzc3N1cmUuZ2V0KCJwZW5kaW5nX3RpY2tzIikgb3IgcHJlc3N1cmUuZ2V0KCJxdWV1ZV9kZXB0aCIpIG9yIDAuMCkKICAgICAgICAgICAgY2FwYWNpdHkgPSBmbG9hdCgKICAgICAgICAgICAgICAgIHByZXNzdXJlLmdldCgicXVldWVfY2FwYWNpdHkiKQogICAgICAgICAgICAgICAgb3IgZ2V0YXR0cihzZWxmLm1kbSwgIl90aWNrX3F1ZXVlX21heHNpemUiLCAwLjApCiAgICAgICAgICAgICAgICBvciAwLjAKICAgICAgICAgICAgKQogICAgICAgICAgICB1dGlsaXphdGlvbiA9IGRlcHRoIC8gY2FwYWNpdHkgaWYgY2FwYWNpdHkgPiAwIGVsc2UgMC4wCiAgICAgICAgICAgIG9wZW5fYWdlID0gZmxvYXQoCiAgICAgICAgICAgICAgICBwcmVzc3N1cmUuZ2V0KCJvbGRlc3Rfb3Blbl9wb3NpdGlvbl9xdW90ZV9hZ2VfcyIpCiAgICAgICAgICAgICAgICBvciBnZXRhdHRyKHNlbGYubWRtLCAib2xkZXN0X29wZW5fcG9zaXRpb25fcXVvdGVfYWdlX3MiLCAwLjApCiAgICAgICAgICAgICAgICBvciAwLjAKICAgICAgICAgICAgKQogICAgICAgICAgICByZXR1cm4gZXZlbnRfbG9vcCwgdGlja19wOTksIHF1ZXVlX2FnZSwgdXRpbGl6YXRpb24sIG9wZW5fYWdlCgogICAgICAgIGRlZiBfc2FtcGxlX2xhdGVuY3koc2VsZikgLT4gTm9uZToKICAgICAgICAgICAgZXZlbnRfbG9vcCwgdGlja19wOTksIHF1ZXVlX2FnZSwgdXRpbGl6YXRpb24sIG9wZW5fYWdlID0gc2VsZi5fbGF0ZW5jeV92YWx1ZXMoKQogICAgICAgICAgICBjcml0aWNhbCA9IGJvb2woCiAgICAgICAgICAgICAgICBldmVudF9sb29wID49IHNlbGYuZXZlbnRfbG9vcF9jcml0aWNhbF9tcwogICAgICAgICAgICAgICAgb3IgdGlja19wOTkgPj0gc2VsZi50aWNrX3A5OV9jcml0aWNhbF9tcwogICAgICAgICAgICAgICAgb3IgcXVldWVfYWdlID49IHNlbGYucXVldWVfYWdlX2NyaXRpY2FsX21zCiAgICAgICAgICAgICAgICBvciB1dGlsaXphdGlvbiA+PSBzZWxmLnF1ZXVlX3V0aWxpemF0aW9uX2NyaXRpY2FsCiAgICAgICAgICAgICAgICBvciBvcGVuX2FnZSA+PSBzZWxmLm9wZW5fcXVvdGVfYWdlX2NyaXRpY2FsX3MKICAgICAgICAgICAgKQogICAgICAgICAgICBpZiBjcml0aWNhbDoKICAgICAgICAgICAgICAgIHNlbGYuX2NyaXRpY2FsX2xhdGVuY3lfc2FtcGxlcyArPSAxCiAgICAgICAgICAgICAgICBzZWxmLl9oZWFsdGh5X2xhdGVuY3lfc2FtcGxlcyA9IDAKICAgICAgICAgICAgZWxzZToKICAgICAgICAgICAgICAgIHNlbGYuX2hlYWx0aHlfbGF0ZW5jeV9zYW1wbGVzICs9IDEKICAgICAgICAgICAgICAgIHNlbGYuX2NyaXRpY2FsX2xhdGVuY3lfc2FtcGxlcyA9IDAKICAgICAgICAgICAgaWYgKAogICAgICAgICAgICAgICAgbm90IHNlbGYuX2xhdGVuY3lfZGVncmFkZWQKICAgICAgICAgICAgICAgIGFuZCBzZWxmLl9jcml0aWNhbF9sYXRlbmN5X3NhbXBsZXMgPj0gc2VsZi5sYXRlbmN5X2NyaXRpY2FsX3NhbXBsZXMKICAgICAgICAgICAgKToKICAgICAgICAgICAgICAgIHNlbGYuX2xhdGVuY3lfZGVncmFkZWQgPSBUcnVlCiAgICAgICAgICAgICAgICBMT0dHRVIuY3JpdGljYWwoCiAgICAgICAgICAgICAgICAgICAgIk1BUktFVF9EQVRBX0xBVEVOQ1lfQ1JJVElDQUwgZXZlbnRfbG9vcF9tcz0lLjFmIHRpY2tfcDk5X21zPSUuMWYgcXVldWVfYWdlX21zPSUuMWYgcXVldWVfdXRpbGl6YXRpb249JS4zZiBvcGVuX3F1b3RlX2FnZV9zPSUuM2YiLAogICAgICAgICAgICAgICAgICAgIGV2ZW50X2xvb3AsCiAgICAgICAgICAgICAgICAgICAgdGlja19wOTksCiAgICAgICAgICAgICAgICAgICAgcXVldWVfYWdlLAogICAgICAgICAgICAgICAgICAgIHV0aWxpemF0aW9uLAogICAgICAgICAgICAgICAgICAgIG9wZW5fYWdlLAogICAgICAgICAgICAgICAgICAgIGV4dHJhPXsiZXZlbnQiOiAiTUFSS0VUX0RBVEFfTEFURU5DWV9DUklUSUNBTCJ9LAogICAgICAgICAgICAgICAgKQogICAg
