+++ scripts/p0p1_chunks/watchdog.part.03
@@ -1,0 +1,1 @@
+ICAgICAgICAgZWxpZiAoCiAgICAgICAgICAgICAgICBzZWxmLl9sYXRlbmN5X2RlZ3JhZGVkCiAgICAgICAgICAgICAgICBhbmQgc2VsZi5faGVhbHRoeV9sYXRlbmN5X3NhbXBsZXMgPj0gc2VsZi5sYXRlbmN5X3JlY292ZXJ5X3NhbXBsZXMKICAgICAgICAgICAgKToKICAgICAgICAgICAgICAgIHNlbGYuX2xhdGVuY3lfZGVncmFkZWQgPSBGYWxzZQogICAgICAgICAgICAgICAgTE9HR0VSLmluZm8oCiAgICAgICAgICAgICAgICAgICAgIk1BUktFVF9EQVRBX0xBVEVOQ1lfUkVDT1ZFUkVEIiwKICAgICAgICAgICAgICAgICAgICBleHRyYT17ImV2ZW50IjogIk1BUktFVF9EQVRBX0xBVEVOQ1lfUkVDT1ZFUkVEIn0sCiAgICAgICAgICAgICAgICApCiAgICAgICAgICAgIHNlbGYubWRtLmxhdGVuY3lfZGVncmFkZWQgPSBzZWxmLl9sYXRlbmN5X2RlZ3JhZGVkCiAgICAgICAgICAgIHNlbGYuX3NuYXBzaG90ID0gU3VwZXJ2aXNvclNuYXBzaG90KAogICAgICAgICAgICAgICAgc3RhdGU9c2VsZi5fc3RhdGUudmFsdWUsCiAgICAgICAgICAgICAgICBsYXRlbmN5X2RlZ3JhZGVkPXNlbGYuX2xhdGVuY3lfZGVncmFkZWQsCiAgICAgICAgICAgICAgICBldmVudF9sb29wX2xhZ19tcz1ldmVudF9sb29wLAogICAgICAgICAgICAgICAgdGlja19wOTlfbXM9dGlja19wOTksCiAgICAgICAgICAgICAgICBxdWV1ZV9hZ2VfbXM9cXVldWVfYWdlLAogICAgICAgICAgICAgICAgcXVldWVfdXRpbGl6YXRpb249dXRpbGl6YXRpb24sCiAgICAgICAgICAgICAgICBvcGVuX3Bvc2l0aW9uX3F1b3RlX2FnZV9zPW9wZW5fYWdlLAogICAgICAgICAgICAgICAgY29uc2VjdXRpdmVfZmFpbHVyZXM9c2VsZi5fY29uc2VjdXRpdmVfZmFpbHVyZXMsCiAgICAgICAgICAgICAgICByZXN0YXJ0X2NvdW50PXNlbGYuX3Jlc3RhcnRfY291bnQsCiAgICAgICAgICAgICAgICBmYXRhbF9zdGFsZV9zYW1wbGVzPXNlbGYuX2ZhdGFsX3N0YWxlX3NhbXBsZXMsCiAgICAgICAgICAgICkKICAgICAgICAgICAgc2VsZi5tZG0ubWFya2V0X2RhdGFfc3VwZXJ2aXNvcl9zbmFwc2hvdCA9IHNlbGYuX3NuYXBzaG90LnRvX2RpY3QoKQoKICAgICAgICBkZWYgX3Jlc3RhcnQoc2VsZiwgbmFtZTogc3RyLCBub3dfbW9ubzogZmxvYXQpIC0+IGJvb2w6CiAgICAgICAgICAgIGlmIG5vd19tb25vIDwgc2VsZi5fbmV4dF9yZXN0YXJ0X21vbm86CiAgICAgICAgICAgICAgICByZXR1cm4gRmFsc2UKICAgICAgICAgICAgaWYgbm90IHNlbGYuX3Jlc3RhcnRfbG9jay5hY3F1aXJlKGJsb2NraW5nPUZhbHNlKToKICAgICAgICAgICAgICAgIHJldHVybiBGYWxzZQogICAgICAgICAgICB0cnk6CiAgICAgICAgICAgICAgICByZXN0YXJ0ID0gZ2V0YXR0cihzZWxmLm1kbSwgbmFtZSwgTm9uZSkKICAgICAgICAgICAgICAgIGlmIG5vdCBjYWxsYWJsZShyZXN0YXJ0KToKICAgICAgICAgICAgICAgICAgICByZXR1cm4gRmFsc2UKICAgICAgICAgICAgICAgIHNlbGYuX3RyYW5zaXRpb24oU3VwZXJ2aXNvclN0YXRlLlJFQ09OTkVDVElORywgbmFtZSkKICAgICAgICAgICAgICAgIHJlc3RhcnQoKQogICAgICAgICAgICAgICAgc2VsZi5fcmVzdGFydF9jb3VudCArPSAxCiAgICAgICAgICAgICAgICBzZWxmLl9jb25zZWN1dGl2ZV9mYWlsdXJlcyArPSAxCiAgICAgICAgICAgICAgICBiYWNrb2ZmID0gbWluKAogICAgICAgICAgICAgICAgICAgIHNlbGYucmVzdGFydF9iYXNlX3MgKiAoMiAqKiBtYXgoc2VsZi5fY29uc2VjdXRpdmVfZmFpbHVyZXMgLSAxLCAwKSksCiAgICAgICAgICAgICAgICAgICAgc2VsZi5yZXN0YXJ0X21heF9zLAogICAgICAgICAgICAgICAgKQogICAgICAgICAgICAgICAgc2VsZi5fbmV4dF9yZXN0YXJ0X21vbm8gPSBub3dfbW9ubyArIGJhY2tvZmYKICAgICAgICAgICAgICAgIHJldHVybiBUcnVlCiAgICAgICAgICAgIGV4Y2VwdCBFeGNlcHRpb24gYXMgZXhjOiAgIyBub3FhOiBCTEUwMDEKICAgICAgICAgICAgICAgIExPR0dFUi5lcnJvcigKICAgICAgICAgICAgICAgICAgICAiTUFSS0VUX0RBVEFfUkVTVEFSVF9GQUlMRUQgY29tcG9uZW50PSVzIGVycm9yPSVzIiwKICAgICAgICAgICAgICAgICAgICBuYW1lLAogICAgICAgICAgICAgICAgICAgIGV4YywKICAgICAgICAgICAgICAgICAgICBleHRyYT17CiAgICAgICAgICAgICAgICAgICAgICAgICJldmVudCI6ICJNQVJLRVRfREFUQV9SRVNUQVJUX0ZBSUxFRCIsCiAgICAgICAgICAgICAgICAgICAgICAgICJjb21wb25lbnQiOiBuYW1lLAogICAgICAgICAgICAgICAgICAgICAgICAiZXJyb3IiOiBzdHIoZXhjKSwKICAgICAgICAgICAgICAgICAgICB9LAogICAgICAgICAgICAgICAgKQogICAgICAgICAgICAgICAgcmV0dXJuIEZhbHNlCiAgICAgICAgICAgIGZpbmFsbHk6CiAgICAgICAgICAgICAgICBzZWxmLl9yZXN0YXJ0X2xvY2sucmVsZWFzZSgpCgogICAgICAgIGRlZiBzYW1wbGVfb25jZShzZWxmLCAqLCBub3dfd2FsbDogZmxvYXQgfCBOb25lID0gTm9uZSwgbm93X21vbm86IGZsb2F0IHwgTm9uZSA9IE5vbmUpIC0+IFN1cGVydmlzb3JTbmFwc2hvdDoKICAgICAgICAgICAgbm93X3dhbGwgPSB0aW1lLnRpbWUoKSBpZiBub3dfd2FsbCBpcyBOb25lIGVsc2Ugbm93X3dhbGwKICAgICAgICAgICAgbm93X21vbm8gPSB0aW1lLm1vbm90b25pYygpIGlmIG5vd19tb25vIGlzIE5vbmUgZWxzZSBub3dfbW9ubwogICAgICAgICAgICBpZiBub3Qgc2VsZi5tYXJrZXRfb3Blbl9nZXR0ZXIoKToKICAgICAgICAgICAgICAgIHNlbGYuX2ZhdGFsX3N0YWxlX3NhbXBsZXMgPSAwCiAgICAgICAgICAgICAgICBzZWxmLl9jb25zZWN1dGl2ZV9mYWlsdXJlcyA9IDAKICAgICAgICAgICAgICAgIHNlbGYuX3RyYW5zaXRpb24oU3VwZXJ2aXNvclN0YXRlLk1BUktFVF9DTE9TRUQsICJtYXJrZXRfY2xvc2VkIikKICAgICAgICAgICAgICAgIHNlbGYuX3NhbXBsZV9sYXRlbmN5KCkKICAgICAgICAgICAgICAgIHJldHVybiBzZWxmLl9zbmFwc2hvdCAgIyB0eXBlOiBpZ25vcmVbcmV0dXJuLXZhbHVlXQoKICAgICAgICAgICAgcG9sbF9hZ2UgPSBfYWdlX3NlY29uZHMoZ2V0YXR0cihzZWxmLm1kbSwgImxhc3RfcG9sbF9oZWFydGJlYXQiLCAwLjApLCBub3dfd2FsbCkKICAgICAgICAgICAgd3NfYWdlID0gX2FnZV9zZWNvbmRzKGdldGF0dHIoc2VsZi5tZG0sICJsYXN0X3dzX2hlYXJ0YmVhdCIsIDAuMCksIG5vd193YWxsKQogICAgICAgICAgICB0aWNrX2FnZSA9IF9hZ2Vfc2Vjb25kcyhnZXRhdHRyKHNlbGYubWRtLCAibGFzdF90aWNrX3RpbWUiLCAwLjApLCBub3dfd2FsbCkKICAgICAgICAgICAgcmVzdGFydGVkID0gRmFsc2UKICAgICAgICAgICAgaWYgcG9sbF9hZ2UgaXMgbm90IE5vbmUgYW5kIHBvbGxfYWdlID49IHNlbGYuc3RhbGVfcmVzdGFydF9zOgogICAgICAgICAgICAgICAgcmVzdGFydGVkID0gc2VsZi5fcmVzdGFydCgicmVzdGFydF9wb2xsaW5nIiwgbm93X21vbm8pIG9yIHJlc3RhcnRlZAogICAgICAgICAgICBpZiB3c19hZ2UgaXMgbm90IE5vbmUgYW5kIHdzX2FnZSA+PSBzZWxmLnN0YWxlX3Jlc3RhcnRfczoKICAgICAgICAgICAgICAgIHJlc3RhcnRlZCA9IHNlbGYuX3Jlc3RhcnQoInJlc3RhcnRfd2Vic29ja2V0Iiwgbm93X21vbm8pIG9yIHJlc3RhcnRlZAoKICAgICAgICAgICAgaWYgdGlja19hZ2UgaXMgbm90IE5vbmUgYW5kIHRpY2tfYWdlID49IHNlbGYuZmF0YWxfc3RhbGVfczoKICAgICAgICAgICAgICAgIHNlbGYuX2ZhdGFsX3N0YWxlX3NhbXBsZXMg
