+++ scripts/p0p1_chunks/watchdog.part.04
@@ -1,0 +1,1 @@
+Kz0gMQogICAgICAgICAgICBlbHNlOgogICAgICAgICAgICAgICAgc2VsZi5fZmF0YWxfc3RhbGVfc2FtcGxlcyA9IDAKICAgICAgICAgICAgICAgIGlmIG5vdCByZXN0YXJ0ZWQ6CiAgICAgICAgICAgICAgICAgICAgc2VsZi5fY29uc2VjdXRpdmVfZmFpbHVyZXMgPSAwCiAgICAgICAgICAgIGlmIHNlbGYuX2ZhdGFsX3N0YWxlX3NhbXBsZXMgPj0gc2VsZi5mYXRhbF9jb25maXJtX3NhbXBsZXM6CiAgICAgICAgICAgICAgICBzZWxmLl90cmFuc2l0aW9uKFN1cGVydmlzb3JTdGF0ZS5GQUlMRUQsICJmYXRhbF90aWNrX3N0YWxlbmVzcyIpCiAgICAgICAgICAgICAgICBMT0dHRVIuY3JpdGljYWwoCiAgICAgICAgICAgICAgICAgICAgIk1BUktFVF9EQVRBX0ZBVEFMX1NUQUxFIHRpY2tfYWdlX3M9JS4xZiBzYW1wbGVzPSVzIGV4aXRpbmc9MSIsCiAgICAgICAgICAgICAgICAgICAgdGlja19hZ2UsCiAgICAgICAgICAgICAgICAgICAgc2VsZi5fZmF0YWxfc3RhbGVfc2FtcGxlcywKICAgICAgICAgICAgICAgICAgICBleHRyYT17ImV2ZW50IjogIk1BUktFVF9EQVRBX0ZBVEFMX1NUQUxFIn0sCiAgICAgICAgICAgICAgICApCiAgICAgICAgICAgICAgICBzZWxmLmV4aXRfcHJvY2VzcygxKQogICAgICAgICAgICBlbGlmIHJlc3RhcnRlZDoKICAgICAgICAgICAgICAgIHNlbGYuX3RyYW5zaXRpb24oU3VwZXJ2aXNvclN0YXRlLlJFQ09OTkVDVElORywgInN0YWxlX3RyYW5zcG9ydCIpCiAgICAgICAgICAgIGVsaWYgYW55KGFnZSBpcyBub3QgTm9uZSBhbmQgYWdlID49IHNlbGYuc3RhbGVfcmVzdGFydF9zIGZvciBhZ2UgaW4gKHBvbGxfYWdlLCB3c19hZ2UsIHRpY2tfYWdlKSk6CiAgICAgICAgICAgICAgICBzZWxmLl90cmFuc2l0aW9uKFN1cGVydmlzb3JTdGF0ZS5ERUdSQURFRCwgInN0YWxlX3RyYW5zcG9ydCIpCiAgICAgICAgICAgIGVsc2U6CiAgICAgICAgICAgICAgICBzZWxmLl90cmFuc2l0aW9uKFN1cGVydmlzb3JTdGF0ZS5IRUFMVEhZLCAidHJhbnNwb3J0X2ZyZXNoIikKICAgICAgICAgICAgc2VsZi5fc2FtcGxlX2xhdGVuY3koKQogICAgICAgICAgICByZXR1cm4gc2VsZi5fc25hcHNob3QgICMgdHlwZTogaWdub3JlW3JldHVybi12YWx1ZV0KCiAgICAgICAgZGVmIF9ydW4oc2VsZikgLT4gTm9uZToKICAgICAgICAgICAgTE9HR0VSLmluZm8oIk1BUktFVF9EQVRBX1NVUEVSVklTT1JfU1RBUlRFRCIpCiAgICAgICAgICAgIHdoaWxlIG5vdCBzZWxmLl9zdG9wLndhaXQoc2VsZi5pbnRlcnZhbF9zKToKICAgICAgICAgICAgICAgIHRyeToKICAgICAgICAgICAgICAgICAgICBzZWxmLnNhbXBsZV9vbmNlKCkKICAgICAgICAgICAgICAgIGV4Y2VwdCBFeGNlcHRpb24gYXMgZXhjOiAgIyBub3FhOiBCTEUwMDEKICAgICAgICAgICAgICAgICAgICBMT0dHRVIuZXJyb3IoCiAgICAgICAgICAgICAgICAgICAgICAgICJNQVJLRVRfREFUQV9TVVBFUlZJU09SX1NBTVBMRV9GQUlMRUQgZXJyb3I9JXMiLAogICAgICAgICAgICAgICAgICAgICAgICBleGMsCiAgICAgICAgICAgICAgICAgICAgICAgIGV4dHJhPXsKICAgICAgICAgICAgICAgICAgICAgICAgICAgICJldmVudCI6ICJNQVJLRVRfREFUQV9TVVBFUlZJU09SX1NBTVBMRV9GQUlMRUQiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgImVycm9yIjogc3RyKGV4YyksCiAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgKQogICAgICAgICAgICBzZWxmLl90cmFuc2l0aW9uKFN1cGVydmlzb3JTdGF0ZS5TVE9QUEVELCAic3RvcF9yZXF1ZXN0ZWQiKQoKCiAgICBkZWYgc3RhcnRfd2F0Y2hkb2cobWFya2V0X2RhdGFfbWFuYWdlcjogQW55KSAtPiB0aHJlYWRpbmcuVGhyZWFkOgogICAgICAgICIiIlN0YXJ0IG9yIHJldXNlIHRoZSBzb2xlIG1hcmtldC1kYXRhIHN1cGVydmlzb3IgdGhyZWFkLiIiIgogICAgICAgIGV4aXN0aW5nID0gZ2V0YXR0cihtYXJrZXRfZGF0YV9tYW5hZ2VyLCAiX21hcmtldF9kYXRhX3N1cGVydmlzb3IiLCBOb25lKQogICAgICAgIGlmIGlzaW5zdGFuY2UoZXhpc3RpbmcsIE1hcmtldERhdGFTdXBlcnZpc29yKToKICAgICAgICAgICAgcmV0dXJuIGV4aXN0aW5nLnN0YXJ0KCkKICAgICAgICBzdXBlcnZpc29yID0gTWFya2V0RGF0YVN1cGVydmlzb3IobWFya2V0X2RhdGFfbWFuYWdlcikKICAgICAgICBtYXJrZXRfZGF0YV9tYW5hZ2VyLl9tYXJrZXRfZGF0YV9zdXBlcnZpc29yID0gc3VwZXJ2aXNvcgogICAgICAgIHJldHVybiBzdXBlcnZpc29yLnN0YXJ0KCkKCgogICAgZGVmIHN0b3Bfd2F0Y2hkb2cobWFya2V0X2RhdGFfbWFuYWdlcjogQW55LCB0aW1lb3V0OiBmbG9hdCA9IDMuMCkgLT4gYm9vbDoKICAgICAgICBzdXBlcnZpc29yID0gZ2V0YXR0cihtYXJrZXRfZGF0YV9tYW5hZ2VyLCAiX21hcmtldF9kYXRhX3N1cGVydmlzb3IiLCBOb25lKQogICAgICAgIGlmIG5vdCBpc2luc3RhbmNlKHN1cGVydmlzb3IsIE1hcmtldERhdGFTdXBlcnZpc29yKToKICAgICAgICAgICAgcmV0dXJuIFRydWUKICAgICAgICByZXR1cm4gc3VwZXJ2aXNvci5zdG9wKHRpbWVvdXQ9dGltZW91dCkKCgogICAgX19hbGxfXyA9IFsKICAgICAgICAiTWFya2V0RGF0YVN1cGVydmlzb3IiLAogICAgICAgICJTdXBlcnZpc29yU25hcHNob3QiLAogICAgICAgICJTdXBlcnZpc29yU3RhdGUiLAogICAgICAgICJzdGFydF93YXRjaGRvZyIsCiAgICAgICAgInN0b3Bfd2F0Y2hkb2ciLAogICAgXQonJycpCndyaXRlKCJzcmMvbmlmdHlfc2NhbHBlcl9ib3QvaW5mcmEvd2F0Y2hkb2cucHkiLCB3YXRjaGRvZykKCg==
