+++ scripts/p0p1_chunks/worker.part.00
@@ -1,0 +1,1 @@
+ZnJvbSBfX2Z1dHVyZV9fIGltcG9ydCBhbm5vdGF0aW9ucwoKaW1wb3J0IHJlCmZyb20gcGF0aGxpYiBpbXBvcnQgUGF0aApmcm9tIHRleHR3cmFwIGltcG9ydCBkZWRlbnQKClJPT1QgPSBQYXRoLmN3ZCgpCgoKZGVmIHJlYWQocGF0aDogc3RyKSAtPiBzdHI6CiAgICByZXR1cm4gKFJPT1QgLyBwYXRoKS5yZWFkX3RleHQoZW5jb2Rpbmc9InV0Zi04IikKCgpkZWYgd3JpdGUocGF0aDogc3RyLCB0ZXh0OiBzdHIpIC0+IE5vbmU6CiAgICB0YXJnZXQgPSBST09UIC8gcGF0aAogICAgdGFyZ2V0LnBhcmVudC5ta2RpcihwYXJlbnRzPVRydWUsIGV4aXN0X29rPVRydWUpCiAgICB0YXJnZXQud3JpdGVfdGV4dCh0ZXh0LCBlbmNvZGluZz0idXRmLTgiKQoKCmRlZiByZXBsYWNlX29uY2UodGV4dDogc3RyLCBvbGQ6IHN0ciwgbmV3OiBzdHIsICosIGxhYmVsOiBzdHIpIC0+IHN0cjoKICAgIGNvdW50ID0gdGV4dC5jb3VudChvbGQpCiAgICBpZiBjb3VudCAhPSAxOgogICAgICAgIHJhaXNlIFJ1bnRpbWVFcnJvcihmIntsYWJlbH06IGV4cGVjdGVkIGV4YWN0bHkgb25lIG1hdGNoLCBmb3VuZCB7Y291bnR9IikKICAgIHJldHVybiB0ZXh0LnJlcGxhY2Uob2xkLCBuZXcsIDEpCgoKZGVmIHJlZ2V4X29uY2UodGV4dDogc3RyLCBwYXR0ZXJuOiBzdHIsIHJlcGxhY2VtZW50OiBzdHIsICosIGxhYmVsOiBzdHIsIGZsYWdzOiBpbnQgPSAwKSAtPiBzdHI6CiAgICB1cGRhdGVkLCBjb3VudCA9IHJlLnN1Ym4ocGF0dGVybiwgcmVwbGFjZW1lbnQsIHRleHQsIGNvdW50PTEsIGZsYWdzPWZsYWdzKQogICAgaWYgY291bnQgIT0gMToKICAgICAgICByYWlzZSBSdW50aW1lRXJyb3IoZiJ7bGFiZWx9OiBleHBlY3RlZCBleGFjdGx5IG9uZSByZWdleCBtYXRjaCwgZm91bmQge2NvdW50fSIpCiAgICByZXR1cm4gdXBkYXRlZAoKCiMgLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tCiMgUDA6IGZhbGxiYWNrIHRpY2sgd29ya2VyIG11c3QgaGF2ZSBvbmUgb3duZXIgYW5kIGZhaWwgY2xvc2VkIG9uIHN0b3AgdGltZW91dC4KIyAtLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0KcGF0aCA9ICJzcmMvbmlmdHlfc2NhbHBlcl9ib3QvZGF0YS9tYXJrZXRfZGF0YV9oYXJkZW5pbmcucHkiCnRleHQgPSByZWFkKHBhdGgpCnRleHQgPSByZXBsYWNlX29uY2UoCiAgICB0ZXh0LAogICAgIiIiICAgIGlmIG5vdCBpc2luc3RhbmNlKGdldGF0dHIoc2VsZiwgXCJfdGlja193b3JrZXJfc3RvcFwiLCBOb25lKSwgdGhyZWFkaW5nLkV2ZW50KTpcbiAgICAgICAgc2VsZi5fdGlja193b3JrZXJfc3RvcCA9IHRocmVhZGluZy5FdmVudCgpXG4gICAgc2VsZi5fY2FuZGxlX2ZsdXNoX3Rhc2s6IGFzeW5jaW8uVGFza1tOb25lXSB8IE5vbmUgPSBOb25lXG4iIiIsCiAgICAiIiIgICAgaWYgbm90IGlzaW5zdGFuY2UoZ2V0YXR0cihzZWxmLCBcIl90aWNrX3dvcmtlcl9zdG9wXCIsIE5vbmUpLCB0aHJlYWRpbmcuRXZlbnQpOlxuICAgICAgICBzZWxmLl90aWNrX3dvcmtlcl9zdG9wID0gdGhyZWFkaW5nLkV2ZW50KClcbiAgICBzZWxmLl90aWNrX3dvcmtlcl9saWZlY3ljbGVfbG9jayA9IHRocmVhZGluZy5STG9jaygpXG4gICAgc2VsZi5fdGlja193b3JrZXJfc3RhdGUgPSBcIlNUT1BQRURcIlxuICAgIHNlbGYuX3RpY2tfd29ya2VyX2dlbmVyYXRpb24gPSAwXG4gICAgc2VsZi5fdGlja193b3JrZXJfZmFpbGVkX3RvX3N0b3AgPSBGYWxzZVxuICAgIHNlbGYuX3RpY2tfd29ya2VyX3N0b3BfdGltZW91dF9zID0gX2Zsb2F0X2VudihcbiAgICAgICAgXCJNRE1fVElDS19XT1JLRVJfU1RPUF9USU1FT1VUX1NFQ09ORFNcIiwgMi4wLCBtaW5pbXVtPTAuMDVcbiAgICApXG4gICAgc2VsZi5fY2FuZGxlX2ZsdXNoX3Rhc2s6IGFzeW5jaW8uVGFza1tOb25lXSB8IE5vbmUgPSBOb25lXG4iIiIsCiAgICBsYWJlbD0id29ya2VyIGxpZmVjeWNsZSBpbml0aWFsaXphdGlvbiIsCikKdGV4dCA9IHJlcGxhY2Vfb25jZSgKICAgIHRleHQsCiAgICAiIiIgICAgc2VsZi5fZW5zdXJlX3RpY2tfd29ya2VyKClcbiAgICBpZiBub3Qgc2VsZi5fcHV0X2ZhbGxiYWNrX3RpY2tfbm93YWl0KHBheWxvYWQpOlxuIiIiLAogICAgIiIiICAgIGlmIG5vdCBzZWxmLl9lbnN1cmVfdGlja193b3JrZXIoKTpcbiAgICAgICAgc2VsZi5fdGlja19xdWV1ZV9kcm9wcGVkICs9IDFcbiAgICAgICAgc2VsZi5fcmVjb3JkX3RpY2tfZHJvcChcbiAgICAgICAgICAgIHN0cihwYXlsb2FkLmdldChcIl9tZG1fcHJpb3JpdHlfYnVja2V0XCIpIG9yIFwidW5rbm93blwiKSxcbiAgICAgICAgICAgIFwiZmFsbGJhY2tfd29ya2VyX3VuYXZhaWxhYmxlXCIsXG4gICAgICAgIClcbiAgICAgICAgc2VsZi5fbG9nZ2VyLmVycm9yKFxuICAgICAgICAgICAgXCJNRE1fRkFMTEJBQ0tfVElDS19SRUpFQ1RFRCByZWFzb249d29ya2VyX3VuYXZhaWxhYmxlIHN0YXRlPSVzXCIsXG4gICAgICAgICAgICBnZXRhdHRyKHNlbGYsIFwiX3RpY2tfd29ya2VyX3N0YXRlXCIsIFwidW5rbm93blwiKSxcbiAgICAgICAgICAgIGV4dHJhPXtcbiAgICAgICAgICAgICAgICBcImV2ZW50XCI6IFwiTURNX0ZBTExCQUNLX1RJQ0tfUkVKRUNURURcIixcbiAgICAgICAgICAgICAgICBcInJlYXNvblwiOiBcIndvcmtlcl91bmF2YWlsYWJsZVwiLFxuICAgICAgICAgICAgICAgIFwid29ya2VyX3N0YXRlXCI6IGdldGF0dHIoc2VsZiwgXCJfdGlja193b3JrZXJfc3RhdGVcIiwgXCJ1bmtub3duXCIpLFxuICAgICAgICAgICAgfSxcbiAgICAgICAgKVxuICAgICAgICByZXR1cm5cbiAgICBpZiBub3Qgc2VsZi5fcHV0X2ZhbGxiYWNrX3RpY2tfbm93YWl0KHBheWxvYWQpOlxuIiIiLAogICAgbGFiZWw9IndvcmtlciBhdmFpbGFiaWxpdHkgZ2F0ZSIsCikKdGV4dCA9IHJlZ2V4X29uY2UoCiAgICB0ZXh0LAogICAgciJkZWYgX2Vuc3VyZV90aWNrX3dvcmtlcl90aHJlYWRfcXVldWVcKHNlbGY6IEFueVwpIC0+IE5vbmU6XG4uKj9cblxuZGVmIF9zYWZlX3F1ZXVlX3Rhc2tfZG9uZSIsCiAgICBkZWRlbnQoCiAgICAgICAgJycnXAogICAgICAgIGRlZiBfZW5zdXJlX3RpY2tfd29ya2VyX3RocmVhZF9xdWV1ZShzZWxmOiBBbnkpIC0+IGJvb2w6CiAgICAgICAgICAgICIiIlN0YXJ0IGV4YWN0bHkgb25lIGZhbGxiYWNrIHdvcmtlcjsgbmV2ZXIgcmVwbGFjZSBhIGxpdmUgb3duZXIuIiIiCiAgICAgICAgICAgIGlmIG5vdCBpc2luc3RhbmNlKGdldGF0dHIoc2VsZiwgIl9mYWxsYmFja190aWNrX3F1ZXVlIiwgTm9uZSksIHF1ZXVlLlF1ZXVlKToKICAgICAgICAgICAgICAgIG1heHNpemUgPSBtYXgoCiAgICAgICAgICAgICAgICAgICAgaW50KGdldGF0dHIoc2VsZiwgIl90aWNrX3F1ZXVlX21heHNpemUiLCAxMF8wMDApIG9yIDEwXzAwMCksIDEKICAgICAgICAgICAgICAgICkKICAgICAgICAgICAgICAgIHNlbGYuX2ZhbGxiYWNrX3RpY2tfcXVldWUgPSBxdWV1ZS5RdWV1ZShtYXhzaXplPW1heHNpemUpCiAgICAgICAgICAgIGxpZmVjeWNsZV9sb2NrID0gZ2V0YXR0cihzZWxmLCAiX3RpY2tfd29ya2VyX2xpZmVjeWNsZV9sb2NrIiwgTm9uZSkKICAgICAgICAgICAgaWYgbGlmZWN5Y2xlX2xvY2sgaXMgTm9uZToKICAgICAgICAgICAgICAgIGxpZmVjeWNsZV9sb2NrID0gdGhyZWFkaW5nLlJMb2NrKCkKICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29y
