+++ scripts/p0p1_chunks/worker.part.01
@@ -1,0 +1,1 @@
+a2VyX2xpZmVjeWNsZV9sb2NrID0gbGlmZWN5Y2xlX2xvY2sKICAgICAgICAgICAgd2l0aCBsaWZlY3ljbGVfbG9jazoKICAgICAgICAgICAgICAgIHRocmVhZCA9IGdldGF0dHIoc2VsZiwgIl90aWNrX3dvcmtlcl90aHJlYWQiLCBOb25lKQogICAgICAgICAgICAgICAgaWYgdGhyZWFkIGlzIG5vdCBOb25lIGFuZCB0aHJlYWQuaXNfYWxpdmUoKToKICAgICAgICAgICAgICAgICAgICByZXR1cm4gZ2V0YXR0cihzZWxmLCAiX3RpY2tfd29ya2VyX3N0YXRlIiwgIlJVTk5JTkciKSA9PSAiUlVOTklORyIKICAgICAgICAgICAgICAgIGlmIGdldGF0dHIoc2VsZiwgIl90aWNrX3dvcmtlcl9zdGF0ZSIsICJTVE9QUEVEIikgaW4gewogICAgICAgICAgICAgICAgICAgICJTVE9QUElORyIsCiAgICAgICAgICAgICAgICAgICAgIkZBSUxFRF9UT19TVE9QIiwKICAgICAgICAgICAgICAgIH06CiAgICAgICAgICAgICAgICAgICAgcmV0dXJuIEZhbHNlCiAgICAgICAgICAgICAgICBzZWxmLl90aWNrX3dvcmtlcl9zdG9wLmNsZWFyKCkKICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29ya2VyX2dlbmVyYXRpb24gPSBpbnQoCiAgICAgICAgICAgICAgICAgICAgZ2V0YXR0cihzZWxmLCAiX3RpY2tfd29ya2VyX2dlbmVyYXRpb24iLCAwKSBvciAwCiAgICAgICAgICAgICAgICApICsgMQogICAgICAgICAgICAgICAgZ2VuZXJhdGlvbiA9IHNlbGYuX3RpY2tfd29ya2VyX2dlbmVyYXRpb24KICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29ya2VyX3N0YXRlID0gIlJVTk5JTkciCiAgICAgICAgICAgICAgICBzZWxmLl90aWNrX3dvcmtlcl9mYWlsZWRfdG9fc3RvcCA9IEZhbHNlCiAgICAgICAgICAgICAgICB0aHJlYWQgPSB0aHJlYWRpbmcuVGhyZWFkKAogICAgICAgICAgICAgICAgICAgIHRhcmdldD1zZWxmLl90aWNrX3dvcmtlcl9sb29wLAogICAgICAgICAgICAgICAgICAgIG5hbWU9ZiJtZG0tdGljay13b3JrZXIte2dlbmVyYXRpb259IiwKICAgICAgICAgICAgICAgICAgICBkYWVtb249VHJ1ZSwKICAgICAgICAgICAgICAgICkKICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29ya2VyX3RocmVhZCA9IHRocmVhZAogICAgICAgICAgICAgICAgdGhyZWFkLnN0YXJ0KCkKICAgICAgICAgICAgc2VsZi5fbG9nZ2VyLmluZm8oCiAgICAgICAgICAgICAgICAiTURNX0ZBTExCQUNLX1RJQ0tfV09SS0VSX1NUQVJURUQgcXVldWVfdHlwZT10aHJlYWRpbmcgZ2VuZXJhdGlvbj0lcyIsCiAgICAgICAgICAgICAgICBnZW5lcmF0aW9uLAogICAgICAgICAgICAgICAgZXh0cmE9ewogICAgICAgICAgICAgICAgICAgICJldmVudCI6ICJNRE1fRkFMTEJBQ0tfVElDS19XT1JLRVJfU1RBUlRFRCIsCiAgICAgICAgICAgICAgICAgICAgInF1ZXVlX3R5cGUiOiAicXVldWUuUXVldWUiLAogICAgICAgICAgICAgICAgICAgICJnZW5lcmF0aW9uIjogZ2VuZXJhdGlvbiwKICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICkKICAgICAgICAgICAgcmV0dXJuIFRydWUKCgogICAgICAgIGRlZiBfc2FmZV9xdWV1ZV90YXNrX2RvbmUnJycKICAgICksCiAgICBsYWJlbD0icmVwbGFjZSB3b3JrZXIgZW5zdXJlIiwKICAgIGZsYWdzPXJlLlMsCikKdGV4dCA9IHJlZ2V4X29uY2UoCiAgICB0ZXh0LAogICAgciJkZWYgX3RpY2tfd29ya2VyX2xvb3BfdGhyZWFkX3F1ZXVlXChzZWxmOiBBbnlcKSAtPiBOb25lOlxuLio/XG5cbmRlZiBfc3RvcF9mYWxsYmFja190aWNrX3dvcmtlciIsCiAgICBkZWRlbnQoCiAgICAgICAgJycnXAogICAgICAgIGRlZiBfdGlja193b3JrZXJfbG9vcF90aHJlYWRfcXVldWUoc2VsZjogQW55KSAtPiBOb25lOgogICAgICAgICAgICAiIiJGYWxsYmFjayBzZXJpYWwgdGljayB3b3JrZXIgdXNlZCBvbmx5IGJlZm9yZSB0aGUgYXN5bmNpbyBsb29wIGlzIHdpcmVkLiIiIgogICAgICAgICAgICBxOiBxdWV1ZS5RdWV1ZVtkaWN0W3N0ciwgQW55XV0gPSBzZWxmLl9mYWxsYmFja190aWNrX3F1ZXVlCiAgICAgICAgICAgIGN1cnJlbnQgPSB0aHJlYWRpbmcuY3VycmVudF90aHJlYWQoKQogICAgICAgICAgICB0cnk6CiAgICAgICAgICAgICAgICB3aGlsZSBub3Qgc2VsZi5fdGlja193b3JrZXJfc3RvcC5pc19zZXQoKToKICAgICAgICAgICAgICAgICAgICB0cnk6CiAgICAgICAgICAgICAgICAgICAgICAgIHJhdyA9IHEuZ2V0KHRpbWVvdXQ9MC4yNSkKICAgICAgICAgICAgICAgICAgICBleGNlcHQgcXVldWUuRW1wdHk6CiAgICAgICAgICAgICAgICAgICAgICAgIGNvbnRpbnVlCiAgICAgICAgICAgICAgICAgICAgdHJ5OgogICAgICAgICAgICAgICAgICAgICAgICBzZWxmLl9wcm9jZXNzX3F1ZXVlZF90aWNrKHJhdykKICAgICAgICAgICAgICAgICAgICAgICAgc2VsZi5fdGlja19wcm9jZXNzZWRfdG90YWwgKz0gMQogICAgICAgICAgICAgICAgICAgIGV4Y2VwdCBFeGNlcHRpb24gYXMgZXhjOiAgIyBub3FhOiBCTEUwMDEKICAgICAgICAgICAgICAgICAgICAgICAgc2VsZi5fcmVjb3JkX3RpY2tfZHJvcCgKICAgICAgICAgICAgICAgICAgICAgICAgICAgIHN0cihyYXcuZ2V0KCJfbWRtX3ByaW9yaXR5X2J1Y2tldCIpIG9yICJ1bmtub3duIiksCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZmFsbGJhY2tfcHJvY2Vzc19lcnJvciIsCiAgICAgICAgICAgICAgICAgICAgICAgICkKICAgICAgICAgICAgICAgICAgICAgICAgc2VsZi5fbG9nZ2VyLmVycm9yKAogICAgICAgICAgICAgICAgICAgICAgICAgICAgIk1ETV9GQUxMQkFDS19USUNLX1dPUktFUl9FUlJPUiBlcnJvcj0lciIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICBleGMsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICBleGNfaW5mbz1UcnVlLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgZXh0cmE9ewogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJldmVudCI6ICJNRE1fRkFMTEJBQ0tfVElDS19XT1JLRVJfRVJST1IiLAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICJlcnJvciI6IHJlcHIoZXhjKSwKICAgICAgICAgICAgICAgICAgICAgICAgICAgIH0sCiAgICAgICAgICAgICAgICAgICAgICAgICkKICAgICAgICAgICAgICAgICAgICBmaW5hbGx5OgogICAgICAgICAgICAgICAgICAgICAgICBfc2FmZV9xdWV1ZV90YXNrX2RvbmUocSkKICAgICAgICAgICAgZmluYWxseToKICAgICAgICAgICAgICAgIGxpZmVjeWNsZV9sb2NrID0gZ2V0YXR0cigKICAgICAgICAgICAgICAgICAgICBzZWxmLCAiX3RpY2tfd29ya2VyX2xpZmVjeWNsZV9sb2NrIiwgdGhyZWFkaW5nLlJMb2NrKCkKICAgICAgICAgICAgICAgICkKICAgICAgICAgICAgICAgIHdpdGggbGlmZWN5Y2xlX2xvY2s6CiAgICAgICAgICAgICAgICAgICAgaWYgZ2V0YXR0cihzZWxmLCAiX3RpY2tfd29ya2VyX3RocmVhZCIsIE5vbmUpIGlzIGN1cnJlbnQ6CiAgICAgICAgICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29ya2VyX3RocmVhZCA9IE5vbmUKICAgICAgICAgICAgICAgICAgICAgICAgc2VsZi5fdGlja193b3JrZXJfc3RhdGUgPSAiU1RPUFBFRCIKICAgICAgICAgICAgICAgICAgICAgICAgc2VsZi5fdGlja193b3JrZXJfZmFpbGVkX3RvX3N0b3AgPSBGYWxzZQoKCiAgICAgICAgZGVmIF9zdG9wX2ZhbGxiYWNrX3RpY2tfd29ya2VyJycnCiAgICApLAogICAgbGFiZWw9InJlcGxhY2Ugd29ya2VyIGxvb3AiLAogICAgZmxhZ3M9cmUuUywKKQp0ZXh0ID0gcmVnZXhfb25jZSgKICAgIHRleHQsCiAgICByImRlZiBfc3Rv
