+++ scripts/p0p1_chunks/worker.part.02
@@ -1,0 +1,1 @@
+cF9mYWxsYmFja190aWNrX3dvcmtlclwoc2VsZjogQW55XCkgLT4gTm9uZTpcbi4qP1xuXG5kZWYgX2Vuc3VyZV9jYW5kbGVfZmx1c2hfdGFzayIsCiAgICBkZWRlbnQoCiAgICAgICAgJycnXAogICAgICAgIGRlZiBfc3RvcF9mYWxsYmFja190aWNrX3dvcmtlcihzZWxmOiBBbnkpIC0+IGJvb2w6CiAgICAgICAgICAgICIiIlN0b3AgdGhlIGN1cnJlbnQgb3duZXIgd2l0aG91dCBtYWtpbmcgYSB0aW1lZC1vdXQgd29ya2VyIHJlc3RhcnRhYmxlLiIiIgogICAgICAgICAgICBsaWZlY3ljbGVfbG9jayA9IGdldGF0dHIoc2VsZiwgIl90aWNrX3dvcmtlcl9saWZlY3ljbGVfbG9jayIsIE5vbmUpCiAgICAgICAgICAgIGlmIGxpZmVjeWNsZV9sb2NrIGlzIE5vbmU6CiAgICAgICAgICAgICAgICBsaWZlY3ljbGVfbG9jayA9IHRocmVhZGluZy5STG9jaygpCiAgICAgICAgICAgICAgICBzZWxmLl90aWNrX3dvcmtlcl9saWZlY3ljbGVfbG9jayA9IGxpZmVjeWNsZV9sb2NrCiAgICAgICAgICAgIHdpdGggbGlmZWN5Y2xlX2xvY2s6CiAgICAgICAgICAgICAgICB0aHJlYWQgPSBnZXRhdHRyKHNlbGYsICJfdGlja193b3JrZXJfdGhyZWFkIiwgTm9uZSkKICAgICAgICAgICAgICAgIGlmIHRocmVhZCBpcyBOb25lOgogICAgICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29ya2VyX3N0YXRlID0gIlNUT1BQRUQiCiAgICAgICAgICAgICAgICAgICAgcmV0dXJuIFRydWUKICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29ya2VyX3N0YXRlID0gIlNUT1BQSU5HIgogICAgICAgICAgICAgICAgc2VsZi5fdGlja193b3JrZXJfc3RvcC5zZXQoKQogICAgICAgICAgICBpZiB0aHJlYWQgaXMgdGhyZWFkaW5nLmN1cnJlbnRfdGhyZWFkKCk6CiAgICAgICAgICAgICAgICByZXR1cm4gRmFsc2UKICAgICAgICAgICAgdGltZW91dCA9IG1heCgKICAgICAgICAgICAgICAgIGZsb2F0KGdldGF0dHIoc2VsZiwgIl90aWNrX3dvcmtlcl9zdG9wX3RpbWVvdXRfcyIsIDIuMCkgb3IgMi4wKSwKICAgICAgICAgICAgICAgIDAuMDUsCiAgICAgICAgICAgICkKICAgICAgICAgICAgaWYgdGhyZWFkLmlzX2FsaXZlKCk6CiAgICAgICAgICAgICAgICB0aHJlYWQuam9pbih0aW1lb3V0PXRpbWVvdXQpCiAgICAgICAgICAgIHdpdGggbGlmZWN5Y2xlX2xvY2s6CiAgICAgICAgICAgICAgICBpZiB0aHJlYWQuaXNfYWxpdmUoKToKICAgICAgICAgICAgICAgICAgICBzZWxmLl90aWNrX3dvcmtlcl9zdGF0ZSA9ICJGQUlMRURfVE9fU1RPUCIKICAgICAgICAgICAgICAgICAgICBzZWxmLl90aWNrX3dvcmtlcl9mYWlsZWRfdG9fc3RvcCA9IFRydWUKICAgICAgICAgICAgICAgICAgICBzZWxmLl9sb2dnZXIuY3JpdGljYWwoCiAgICAgICAgICAgICAgICAgICAgICAgICJNRE1fRkFMTEJBQ0tfVElDS19XT1JLRVJfU1RPUF9USU1FT1VUIGdlbmVyYXRpb249JXMgdGltZW91dF9zPSUuM2YiLAogICAgICAgICAgICAgICAgICAgICAgICBnZXRhdHRyKHNlbGYsICJfdGlja193b3JrZXJfZ2VuZXJhdGlvbiIsIDApLAogICAgICAgICAgICAgICAgICAgICAgICB0aW1lb3V0LAogICAgICAgICAgICAgICAgICAgICAgICBleHRyYT17CiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZXZlbnQiOiAiTURNX0ZBTExCQUNLX1RJQ0tfV09SS0VSX1NUT1BfVElNRU9VVCIsCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAiZ2VuZXJhdGlvbiI6IGdldGF0dHIoc2VsZiwgIl90aWNrX3dvcmtlcl9nZW5lcmF0aW9uIiwgMCksCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAidGltZW91dF9zIjogdGltZW91dCwKICAgICAgICAgICAgICAgICAgICAgICAgfSwKICAgICAgICAgICAgICAgICAgICApCiAgICAgICAgICAgICAgICAgICAgcmV0dXJuIEZhbHNlCiAgICAgICAgICAgICAgICBpZiBnZXRhdHRyKHNlbGYsICJfdGlja193b3JrZXJfdGhyZWFkIiwgTm9uZSkgaXMgdGhyZWFkOgogICAgICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29ya2VyX3RocmVhZCA9IE5vbmUKICAgICAgICAgICAgICAgIHNlbGYuX3RpY2tfd29ya2VyX3N0YXRlID0gIlNUT1BQRUQiCiAgICAgICAgICAgICAgICBzZWxmLl90aWNrX3dvcmtlcl9mYWlsZWRfdG9fc3RvcCA9IEZhbHNlCiAgICAgICAgICAgICAgICByZXR1cm4gVHJ1ZQoKCiAgICAgICAgZGVmIF9lbnN1cmVfY2FuZGxlX2ZsdXNoX3Rhc2snJycKICAgICksCiAgICBsYWJlbD0icmVwbGFjZSB3b3JrZXIgc3RvcCIsCiAgICBmbGFncz1yZS5TLAopCndyaXRlKHBhdGgsIHRleHQpCgo=
